% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak-list-functions.R
\name{consensusPeaks}
\alias{consensusPeaks}
\title{Reduce mass peaks from input spectra to consensus peaks}
\usage{
consensusPeaks(x, minProp = 0.5, intensityFun = base::mean,
  mzFun = base::mean, weighted = FALSE, tolerance = 0, ppm = 0,
  ...)
}
\arguments{
\item{x}{\code{list} of peak matrices.}

\item{minProp}{\code{numeric(1)} with the minimal required proportion of input
spectra (peak matrices) a mass peak has to be present to be included
in the consensus peak matrix.}

\item{intensityFun}{\code{function} to be used to combine intensity values for
matching peaks. By default the mean intensity value is returned.}

\item{mzFun}{\code{function} to be used to combine m/z values for matching peaks.
By default the mean m/z value is returned.}

\item{weighted}{\code{logical(1)} defining whether m/z values for matching peaks
should be calculated by an intensity-weighted average of the individuak
m/z values. This overrides parameter \code{mzFun}.}

\item{tolerance}{\code{numeric(1)} defining the (absolute) maximal accepted
difference between mass peaks to group them into the same final peak.}

\item{ppm}{\code{numeric(1)} defining the m/z-relative maximal accepted
difference between mass peaks (expressed in parts-per-million) to group
them into the same final peak.}

\item{...}{additional parameters to the \code{mzFun} and \code{intensityFun} functions.}
}
\value{
Peaks \code{matrix} with m/z and intensity values representing the consensus peaks
from the provided peak matrices.
}
\description{
\code{consensusPeaks} reduces the peaks from all input spectra to a set of
\emph{consensus} peaks, i.e. peaks present in a certain proportion of input
spectra (which can be defined with parameter \code{minProp}).

For details on the mass peak grouping across samples see \code{\link[=combinePeaks]{combinePeaks()}}.
}
\examples{

p1 <- cbind(mz = c(12, 45, 64, 70), intensity = c(10, 20, 30, 40))
p2 <- cbind(mz = c(17, 45.1, 63.9, 70.2), intensity = c(11, 21, 31, 41))
p3 <- cbind(mz = c(12.1, 44.9, 63), intensity = c(12, 22, 32))

## No mass peaks identical thus consensus peaks are empty
consensusPeaks(list(p1, p2, p3))

## Reducing the minProp to 0.2. The consensus spectrum will contain all
## peaks
consensusPeaks(list(p1, p2, p3), minProp = 0.2)

## With a tolerance of 0.1 mass peaks can be matched across spectra
consensusPeaks(list(p1, p2, p3), tolerance = 0.1)

## Report the minimal m/z and intensity
consensusPeaks(list(p1, p2, p3), tolerance = 0.1, intensityFun = min,
    mzFun = min)
}
\seealso{
Other peak matrix combining functions: \code{\link{combinePeaks}}
}
\author{
Johannes Rainer
}
\concept{peak matrix combining functions}
