% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dictionary.R
\docType{data}
\name{dictionary}
\alias{dictionary}
\title{Dictionary}
\description{
A dictionary, detailing the meaning of variables listed in \code{\link{meta_data}}.
}
\author{
Thomas White \email{thomas.white026@gmail.com}
}
\keyword{data}
