% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft_spectrum.R
\name{filterFourierTransformArtefacts}
\alias{filterFourierTransformArtefacts}
\title{Fast fourier transform artefact filter}
\description{
The \code{filterFourierTransformArtefacts} function removes (Orbitrap) fast
fourier artefact peaks from spectra. Such artefacts (also referred to as
\emph{rippples}) seem to be related to the
\href{https://en.wikipedia.org/wiki/Ringing_artifacts}{\emph{ringing}} phenomenon and
are frequently seen in Orbitrap data as small random mass peaks ~ 0.01 Da
from a main peak with a very large intensity. See also
\href{https://www.shimadzu.com/an/service-support/technical-support/analysis-basics/tips-ftir/apodization.html}{here}
for more details and information.

See also \code{\link[=Spectra]{Spectra()}} (section *Data subsetting, filtering and merging) for
the definition of the function.
}
\details{
The current implementation iterates through all intensity ordered peaks in a
spectrum and removes all peaks with an m/z within +/- \code{halfWindowSize} of
the current peak if their intensity is lower than \code{threshold} times the
current peak's intensity. Additional parameters \code{keepIsotopes}, \code{maxCharge}
and \code{isotopeTolerance} allow to avoid removing of potential \verb{[13]C} isotope
peaks (\code{maxCharge} being the maximum charge that should be considered
and \code{isotopeTolerance} the absolute acceptable tolerance for matching
their m/z).
}
\author{
Jan Stanstrup, Johannes Rainer
}
