% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks-functions.R
\name{gnps}
\alias{gnps}
\title{GNPS spectra similarity score}
\usage{
gnps(x, y)
}
\arguments{
\item{x}{\code{numeric} peak \code{matrix} with m/z (first column) and intensity
values (second column). Rows in \code{x} and \code{y} are considered to contain
matching peaks, i.e. the first row in \code{x} contains a peak matching the
first peak (row) of \code{y} or \code{NA_real_} if the peak does not match.}

\item{y}{\code{numeric} peak \code{matrix} with m/z (first column) and intensity
values (second column). Rows in \code{x} and \code{y} are considered to contain
matching peaks, i.e. the first row in \code{x} contains a peak matching the
first peak (row) of \code{y} or \code{NA_real_} if the peak does not match.}
}
\description{
The GNPS approach considers in its similarity score, in addition to directly
matching peaks between two spectra, also matches of peaks which difference in
m/z values matches the difference of the spectra's precursor m/z. For peaks
that match multiple peaks in the other spectrum the matching peak pair with
the higher value/similarity is considered in the final similarity score.

To calculate GNPS scores, the \code{\link[=joinPeaksGnps]{joinPeaksGnps()}} function should be used first
to match the peaks between the compared spectra and \code{gnps} should then be
called on the resulting aligned peak matrices. To use the GNPS score on a
\code{\link[=Spectra]{Spectra()}} object \code{compareSpectra} should be used with
\code{MAPFUN = joinPeaksGnps} and \code{FUN = gnps}.
}
\details{
The implementation bases on the R code from the publication listed in the
references.
}
\examples{

## Define spectra
x <- cbind(mz = c(10, 36, 63, 91, 93), intensity = c(14, 15, 999, 650, 1))
y <- cbind(mz = c(10, 12, 50, 63, 105), intensity = c(35, 5, 16, 999, 450))
## The precursor m/z
pmz_x <- 91
pmz_y <- 105

## joinPeaksGnps finds 4 matches
map <- joinPeaksGnps(x, y, pmz_x, pmz_y)
map

## Calculate similarity
gnps(map$x, map$y)
}
\references{
Xing S, Hu Y, Yin Z, Liu M, Tang X, Fang M, Huan T. Retrieving and Utilizing
Hypothetical Neutral Losses from Tandem Mass Spectra for Spectral Similarity
Analysis and Unknown Metabolite Annotation. \emph{Anal Chem.}
2020 Nov 3;92(21):14476-14483. doi: 10.1021/acs.analchem.0c02521.
}
\author{
Johannes Rainer, Michael Witting, based on the code from the ref
}
