% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-meta.R
\docType{data}
\name{meta_data}
\alias{meta_data}
\title{Metadata for reflectance spectra}
\description{
Metadata for all spectra. See \code{\link{dictionary}}
 for details of the variable meanings.
}
\author{
Thomas White \email{thomas.white026@gmail.com}
}
\keyword{data}
