% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-util.R
\name{rbindlistWithRownames}
\alias{rbindlistWithRownames}
\title{Fast \emph{rbind-ing} \code{data.frame}s preserving row names}
\usage{
rbindlistWithRownames(
  l,
  use.names = TRUE,
  fill = FALSE,
  idcol = NULL,
  ignore.attr = FALSE
)
}
\arguments{
\item{l}{A \code{list} containing \code{data.frame}s that should be combined.}

\item{use.names}{If \code{TRUE} binds by matching column names, if \code{FALSE} by
position. \code{"check"} (default) warns if not all items have the same names
in the same order. See \code{\link[data.table:rbindlist]{data.table::rbindlist()}} for details.}

\item{fill}{\code{logical(1)}: if \code{TRUE} fills missing columns with NAs or \code{NULL}
for missing list columns. By defalt \code{FALSE}.}

\item{idcol}{Creates a column in the result showing which list item those
rows cam from. See \code{\link[data.table:rbindlist]{data.table::rbindlist()}} for details.}

\item{ignore.attr}{\code{logical(1)}, default \code{FALSE}. When \code{TRUE}, allows
binding columns with different attributes (e.g. class).}
}
\value{
A combined \code{data.frame} with row names (if present in all \code{data.frame}s
provided).
}
\description{
The \code{rbindlistWithRownames()} function uses the \code{\link[data.table:rbindlist]{data.table::rbindlist()}}
function for a fast concatenation (row-binding) of a \code{list} of provided
\code{data.frame}s and in addition also preserves their row names (if set).

The parameters are the same as for \code{rbindlist()}.
}
\note{
No sanitizing of row names is performed, i.e., after merging duplicated
names might be present.

The function uses \code{.row_names_info()} to guess wheter row names are \emph{really}
set or just the default \code{seq_len(nrow(x))} are used.
}
\seealso{
\code{\link[data.table:rbindlist]{data.table::rbindlist()}}
}
\author{
Johannes Rainer
}
