% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchspec.R
\name{searchspec}
\alias{searchspec}
\title{Search the spectral reflectance database}
\usage{
searchspec(..., aggregate = FALSE, by = NULL)
}
\arguments{
\item{...}{Search conditions to match (note: must use relational operators, e.g.
\code{==}, \code{<=}). Can match against any variables listed in the \code{\link{dictionary}},
namely:
\itemize{
 \item \code{phylum}:
 \item \code{class}:
 \item \code{order}:
 \item \code{family}:
 \item \code{genus}:
 \item \code{species}:
 \item \code{sub}:
 \item \code{patch}:
 \item \code{sex}:
 \item \code{country}:
 \item \code{latitude}:
 \item \code{longitude}:
 \item \code{study_name}:
 \item \code{contact_name}:
 \item \code{contact_email}:
 \item \code{doi}:
}}

\item{aggregate}{aggregate resulting spectra? Defaults to FALSE.}

\item{by}{single categorical variable to aggregate by if aggregate = TRUE}
}
\value{
A data frame of reflectance spectra from binned in 1 nm intervals
 from 300-700 nm, with the first column containing wavelengths ('wl').
}
\description{
Search the database
}
\examples{
\dontrun{

butterflies <- searchspec(order == 'Lepidoptera')

female_butterflies <- searchspec(order == "Lepidoptera", sex == 'f')

}

}
\author{
Thomas White \email{thomas.white@sydney.edu.au}
}
