% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{summaryspec}
\alias{summaryspec}
\title{Summary metadata}
\usage{
summaryspec(specs)
}
\arguments{
\item{specs}{a data.frame of reflectance spectra, the result of \code{\link{searchspec}}.
If no spectra are provided, a species-level summary of the entire database is returned.}
}
\value{
summary metadata
}
\description{
Summary metadata
}
\examples{
\dontrun{
# Summarise entire current database
summaryspec()

# Summarise a searched subset
butterflies <- searchspec(order == "Lepidoptera")
summaryspec(butterflies)
}


}
\author{
Thomas White \email{thomas.white026@gmail.com}
}
