% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{writespec}
\alias{writespec}
\title{Internal function for writing to the database}
\usage{
writespec()
}
\value{
A data frame of reflectance spectra
}
\description{
Internal function for writing to the database
}
\examples{
\dontrun{
# Dichromat
data(flowers)
}

}
\author{
Thomas White \email{thomas.white026@gmail.com}
}
\keyword{internal}
