% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_aliases.R, R/ProcessingStep.R,
%   R/Spectra.R
\docType{methods}
\name{hidden_aliases}
\alias{hidden_aliases}
\alias{ProcessingStep-class}
\alias{show,ProcessingStep-method}
\alias{show,Spectra-method}
\title{Internal page for hidden aliases}
\usage{
\S4method{show}{ProcessingStep}(object)

\S4method{show}{Spectra}(object)
}
\description{
For S4 methods that require a documentation entry but only clutter the index.
}
\keyword{internal}
