% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_aliases.R, R/AllGenerics.R,
%   R/MsBackendDFrame.R, R/MsBackendDataFrame.R, R/MsBackendHdf5Peaks.R,
%   R/MsBackendMzR.R, R/ProcessingStep.R, R/Spectra.R
\docType{methods}
\name{hidden_aliases}
\alias{hidden_aliases}
\alias{[,MsBackendDataFrame-method}
\alias{ppm}
\alias{backendInitialize}
\alias{backendMerge}
\alias{bin}
\alias{bin,numeric-method}
\alias{clean}
\alias{compareSpectra}
\alias{isReadOnly}
\alias{pickPeaks}
\alias{removePeaks}
\alias{replaceList<-}
\alias{selectSpectraVariables}
\alias{setBackend}
\alias{show,MsBackendDFrame-method}
\alias{backendInitialize,MsBackendDFrame-method}
\alias{backendMerge,MsBackendDFrame-method}
\alias{acquisitionNum,MsBackendDFrame-method}
\alias{as.list,MsBackendDFrame-method}
\alias{centroided,MsBackendDFrame-method}
\alias{centroided<-,MsBackendDFrame-method}
\alias{collisionEnergy,MsBackendDFrame-method}
\alias{collisionEnergy<-,MsBackendDFrame-method}
\alias{dataOrigin,MsBackendDFrame-method}
\alias{dataOrigin<-,MsBackendDFrame-method}
\alias{dataStorage,MsBackendDFrame-method}
\alias{dataStorage<-,MsBackendDFrame-method}
\alias{intensity,MsBackendDFrame-method}
\alias{intensity<-,MsBackendDFrame-method}
\alias{ionCount,MsBackendDFrame-method}
\alias{isCentroided,MsBackendDFrame-method}
\alias{isEmpty,MsBackendDFrame-method}
\alias{isolationWindowLowerMz,MsBackendDFrame-method}
\alias{isolationWindowLowerMz<-,MsBackendDFrame-method}
\alias{isolationWindowTargetMz,MsBackendDFrame-method}
\alias{isolationWindowTargetMz<-,MsBackendDFrame-method}
\alias{isolationWindowUpperMz,MsBackendDFrame-method}
\alias{isolationWindowUpperMz<-,MsBackendDFrame-method}
\alias{length,MsBackendDFrame-method}
\alias{lengths,MsBackendDFrame-method}
\alias{msLevel,MsBackendDFrame-method}
\alias{mz,MsBackendDFrame-method}
\alias{mz<-,MsBackendDFrame-method}
\alias{polarity,MsBackendDFrame-method}
\alias{polarity<-,MsBackendDFrame-method}
\alias{precScanNum,MsBackendDFrame-method}
\alias{precursorCharge,MsBackendDFrame-method}
\alias{precursorIntensity,MsBackendDFrame-method}
\alias{precursorMz,MsBackendDFrame-method}
\alias{replaceList<-,MsBackendDFrame-method}
\alias{rtime,MsBackendDFrame-method}
\alias{rtime<-,MsBackendDFrame-method}
\alias{scanIndex,MsBackendDFrame-method}
\alias{selectSpectraVariables,MsBackendDFrame-method}
\alias{smoothed,MsBackendDFrame-method}
\alias{smoothed<-,MsBackendDFrame-method}
\alias{spectraData,MsBackendDFrame-method}
\alias{spectraData<-,MsBackendDFrame-method}
\alias{spectraNames,MsBackendDFrame-method}
\alias{spectraNames<-,MsBackendDFrame-method}
\alias{spectraVariables,MsBackendDFrame-method}
\alias{tic,MsBackendDFrame-method}
\alias{$,MsBackendDFrame-method}
\alias{$<-,MsBackendDFrame-method}
\alias{[,MsBackendDFrame-method}
\alias{split,MsBackendDFrame,ANY-method}
\alias{filterAcquisitionNum,MsBackendDFrame-method}
\alias{filterDataOrigin,MsBackendDFrame-method}
\alias{filterDataStorage,MsBackendDFrame-method}
\alias{filterEmptySpectra,MsBackendDFrame-method}
\alias{filterIsolationWindow,MsBackendDFrame-method}
\alias{filterMsLevel,MsBackendDFrame-method}
\alias{filterPolarity,MsBackendDFrame-method}
\alias{filterPrecursorMz,MsBackendDFrame-method}
\alias{filterPrecursorScan,MsBackendDFrame-method}
\alias{filterRt,MsBackendDFrame-method}
\alias{show,MsBackendDataFrame-method}
\alias{backendInitialize,MsBackendDataFrame-method}
\alias{backendMerge,MsBackendDataFrame-method}
\alias{acquisitionNum,MsBackendDataFrame-method}
\alias{as.list,MsBackendDataFrame-method}
\alias{centroided,MsBackendDataFrame-method}
\alias{centroided<-,MsBackendDataFrame-method}
\alias{collisionEnergy,MsBackendDataFrame-method}
\alias{collisionEnergy<-,MsBackendDataFrame-method}
\alias{dataOrigin,MsBackendDataFrame-method}
\alias{dataOrigin<-,MsBackendDataFrame-method}
\alias{dataStorage,MsBackendDataFrame-method}
\alias{dataStorage<-,MsBackendDataFrame-method}
\alias{intensity,MsBackendDataFrame-method}
\alias{intensity<-,MsBackendDataFrame-method}
\alias{ionCount,MsBackendDataFrame-method}
\alias{isCentroided,MsBackendDataFrame-method}
\alias{isEmpty,MsBackendDataFrame-method}
\alias{isolationWindowLowerMz,MsBackendDataFrame-method}
\alias{isolationWindowLowerMz<-,MsBackendDataFrame-method}
\alias{isolationWindowTargetMz,MsBackendDataFrame-method}
\alias{isolationWindowTargetMz<-,MsBackendDataFrame-method}
\alias{isolationWindowUpperMz,MsBackendDataFrame-method}
\alias{isolationWindowUpperMz<-,MsBackendDataFrame-method}
\alias{length,MsBackendDataFrame-method}
\alias{lengths,MsBackendDataFrame-method}
\alias{msLevel,MsBackendDataFrame-method}
\alias{mz,MsBackendDataFrame-method}
\alias{mz<-,MsBackendDataFrame-method}
\alias{polarity,MsBackendDataFrame-method}
\alias{polarity<-,MsBackendDataFrame-method}
\alias{precScanNum,MsBackendDataFrame-method}
\alias{precursorCharge,MsBackendDataFrame-method}
\alias{precursorIntensity,MsBackendDataFrame-method}
\alias{precursorMz,MsBackendDataFrame-method}
\alias{replaceList<-,MsBackendDataFrame-method}
\alias{rtime,MsBackendDataFrame-method}
\alias{rtime<-,MsBackendDataFrame-method}
\alias{scanIndex,MsBackendDataFrame-method}
\alias{selectSpectraVariables,MsBackendDataFrame-method}
\alias{smoothed,MsBackendDataFrame-method}
\alias{smoothed<-,MsBackendDataFrame-method}
\alias{spectraData,MsBackendDataFrame-method}
\alias{spectraData<-,MsBackendDataFrame-method}
\alias{spectraNames,MsBackendDataFrame-method}
\alias{spectraNames<-,MsBackendDataFrame-method}
\alias{spectraVariables,MsBackendDataFrame-method}
\alias{tic,MsBackendDataFrame-method}
\alias{$,MsBackendDataFrame-method}
\alias{$<-,MsBackendDataFrame-method}
\alias{split,MsBackendDataFrame,ANY-method}
\alias{filterAcquisitionNum,MsBackendDataFrame-method}
\alias{filterDataOrigin,MsBackendDataFrame-method}
\alias{filterDataStorage,MsBackendDataFrame-method}
\alias{filterEmptySpectra,MsBackendDataFrame-method}
\alias{filterIsolationWindow,MsBackendDataFrame-method}
\alias{filterMsLevel,MsBackendDataFrame-method}
\alias{filterPolarity,MsBackendDataFrame-method}
\alias{filterPrecursorMz,MsBackendDataFrame-method}
\alias{filterPrecursorScan,MsBackendDataFrame-method}
\alias{filterRt,MsBackendDataFrame-method}
\alias{backendInitialize,MsBackendHdf5Peaks-method}
\alias{show,MsBackendHdf5Peaks-method}
\alias{as.list,MsBackendHdf5Peaks-method}
\alias{intensity,MsBackendHdf5Peaks-method}
\alias{intensity<-,MsBackendHdf5Peaks-method}
\alias{ionCount,MsBackendHdf5Peaks-method}
\alias{isCentroided,MsBackendHdf5Peaks-method}
\alias{isEmpty,MsBackendHdf5Peaks-method}
\alias{lengths,MsBackendHdf5Peaks-method}
\alias{mz,MsBackendHdf5Peaks-method}
\alias{mz<-,MsBackendHdf5Peaks-method}
\alias{replaceList<-,MsBackendHdf5Peaks-method}
\alias{spectraData,MsBackendHdf5Peaks-method}
\alias{spectraData<-,MsBackendHdf5Peaks-method}
\alias{$<-,MsBackendHdf5Peaks-method}
\alias{[,MsBackendHdf5Peaks-method}
\alias{backendMerge,MsBackendHdf5Peaks-method}
\alias{backendInitialize,MsBackendMzR-method}
\alias{show,MsBackendMzR-method}
\alias{as.list,MsBackendMzR-method}
\alias{intensity,MsBackendMzR-method}
\alias{intensity<-,MsBackendMzR-method}
\alias{ionCount,MsBackendMzR-method}
\alias{isCentroided,MsBackendMzR-method}
\alias{isEmpty,MsBackendMzR-method}
\alias{lengths,MsBackendMzR-method}
\alias{mz,MsBackendMzR-method}
\alias{mz<-,MsBackendMzR-method}
\alias{spectraData,MsBackendMzR-method}
\alias{spectraData<-,MsBackendMzR-method}
\alias{spectraNames,MsBackendMzR-method}
\alias{spectraNames<-,MsBackendMzR-method}
\alias{spectraVariables,MsBackendMzR-method}
\alias{$<-,MsBackendMzR-method}
\alias{ProcessingStep-class}
\alias{show,ProcessingStep-method}
\alias{show,Spectra-method}
\title{Internal page for hidden aliases}
\usage{
backendInitialize(object, ...)

backendMerge(object, ...)

bin(x, ...)

\S4method{bin}{numeric}(
  x,
  y,
  size = 1,
  breaks = seq(floor(min(y)), ceiling(max(y)), by = size),
  FUN = max
)

clean(object, ...)

compareSpectra(x, y, ...)

isReadOnly(object, ...)

pickPeaks(object, ...)

removePeaks(object, t = "min", ...)

replaceList(object) <- value

selectSpectraVariables(object, ...)

setBackend(object, backend, ...)

\S4method{show}{MsBackendDFrame}(object)

\S4method{backendInitialize}{MsBackendDFrame}(object, spectraData, ...)

\S4method{backendMerge}{MsBackendDFrame}(object, ...)

\S4method{acquisitionNum}{MsBackendDFrame}(object)

\S4method{as.list}{MsBackendDFrame}(x)

\S4method{centroided}{MsBackendDFrame}(object)

\S4method{centroided}{MsBackendDFrame}(object) <- value

\S4method{collisionEnergy}{MsBackendDFrame}(object)

\S4method{collisionEnergy}{MsBackendDFrame}(object) <- value

\S4method{dataOrigin}{MsBackendDFrame}(object)

\S4method{dataOrigin}{MsBackendDFrame}(object) <- value

\S4method{dataStorage}{MsBackendDFrame}(object)

\S4method{dataStorage}{MsBackendDFrame}(object) <- value

\S4method{intensity}{MsBackendDFrame}(object)

\S4method{intensity}{MsBackendDFrame}(object) <- value

\S4method{ionCount}{MsBackendDFrame}(object)

\S4method{isCentroided}{MsBackendDFrame}(object, ...)

\S4method{isEmpty}{MsBackendDFrame}(x)

\S4method{isolationWindowLowerMz}{MsBackendDFrame}(object)

\S4method{isolationWindowLowerMz}{MsBackendDFrame}(object) <- value

\S4method{isolationWindowTargetMz}{MsBackendDFrame}(object)

\S4method{isolationWindowTargetMz}{MsBackendDFrame}(object) <- value

\S4method{isolationWindowUpperMz}{MsBackendDFrame}(object)

\S4method{isolationWindowUpperMz}{MsBackendDFrame}(object) <- value

\S4method{length}{MsBackendDFrame}(x)

\S4method{lengths}{MsBackendDFrame}(x, use.names = FALSE)

\S4method{msLevel}{MsBackendDFrame}(object, ...)

\S4method{mz}{MsBackendDFrame}(object)

\S4method{mz}{MsBackendDFrame}(object) <- value

\S4method{polarity}{MsBackendDFrame}(object)

\S4method{polarity}{MsBackendDFrame}(object) <- value

\S4method{precScanNum}{MsBackendDFrame}(object)

\S4method{precursorCharge}{MsBackendDFrame}(object)

\S4method{precursorIntensity}{MsBackendDFrame}(object)

\S4method{precursorMz}{MsBackendDFrame}(object)

\S4method{replaceList}{MsBackendDFrame}(object) <- value

\S4method{rtime}{MsBackendDFrame}(object)

\S4method{rtime}{MsBackendDFrame}(object) <- value

\S4method{scanIndex}{MsBackendDFrame}(object)

\S4method{selectSpectraVariables}{MsBackendDFrame}(object, spectraVariables = spectraVariables(object))

\S4method{smoothed}{MsBackendDFrame}(object)

\S4method{smoothed}{MsBackendDFrame}(object) <- value

\S4method{spectraData}{MsBackendDFrame}(object, columns = spectraVariables(object))

\S4method{spectraData}{MsBackendDFrame}(object) <- value

\S4method{spectraNames}{MsBackendDFrame}(object)

\S4method{spectraNames}{MsBackendDFrame}(object) <- value

\S4method{spectraVariables}{MsBackendDFrame}(object)

\S4method{tic}{MsBackendDFrame}(object, initial = TRUE)

\S4method{$}{MsBackendDFrame}(x, name)

\S4method{$}{MsBackendDFrame}(x, name) <- value

\S4method{[}{MsBackendDFrame}(x, i, j, ..., drop = FALSE)

\S4method{split}{MsBackendDFrame,ANY}(x, f, drop = FALSE, ...)

\S4method{filterAcquisitionNum}{MsBackendDFrame}(
  object,
  n = integer(),
  dataStorage = character(),
  dataOrigin = character()
)

\S4method{filterDataOrigin}{MsBackendDFrame}(object, dataOrigin = character())

\S4method{filterDataStorage}{MsBackendDFrame}(object, dataStorage = character())

\S4method{filterEmptySpectra}{MsBackendDFrame}(object)

\S4method{filterIsolationWindow}{MsBackendDFrame}(object, mz = numeric(), ...)

\S4method{filterMsLevel}{MsBackendDFrame}(object, msLevel = integer())

\S4method{filterPolarity}{MsBackendDFrame}(object, polarity = integer())

\S4method{filterPrecursorMz}{MsBackendDFrame}(object, mz = numeric())

\S4method{filterPrecursorScan}{MsBackendDFrame}(object, acquisitionNum = integer())

\S4method{filterRt}{MsBackendDFrame}(object, rt = numeric(), msLevel. = unique(msLevel(object)))

\S4method{show}{MsBackendDataFrame}(object)

\S4method{backendInitialize}{MsBackendDataFrame}(object, spectraData, ...)

\S4method{backendMerge}{MsBackendDataFrame}(object, ...)

\S4method{acquisitionNum}{MsBackendDataFrame}(object)

\S4method{as.list}{MsBackendDataFrame}(x)

\S4method{centroided}{MsBackendDataFrame}(object)

\S4method{centroided}{MsBackendDataFrame}(object) <- value

\S4method{collisionEnergy}{MsBackendDataFrame}(object)

\S4method{collisionEnergy}{MsBackendDataFrame}(object) <- value

\S4method{dataOrigin}{MsBackendDataFrame}(object)

\S4method{dataOrigin}{MsBackendDataFrame}(object) <- value

\S4method{dataStorage}{MsBackendDataFrame}(object)

\S4method{dataStorage}{MsBackendDataFrame}(object) <- value

\S4method{intensity}{MsBackendDataFrame}(object)

\S4method{intensity}{MsBackendDataFrame}(object) <- value

\S4method{ionCount}{MsBackendDataFrame}(object)

\S4method{isCentroided}{MsBackendDataFrame}(object, ...)

\S4method{isEmpty}{MsBackendDataFrame}(x)

\S4method{isolationWindowLowerMz}{MsBackendDataFrame}(object)

\S4method{isolationWindowLowerMz}{MsBackendDataFrame}(object) <- value

\S4method{isolationWindowTargetMz}{MsBackendDataFrame}(object)

\S4method{isolationWindowTargetMz}{MsBackendDataFrame}(object) <- value

\S4method{isolationWindowUpperMz}{MsBackendDataFrame}(object)

\S4method{isolationWindowUpperMz}{MsBackendDataFrame}(object) <- value

\S4method{length}{MsBackendDataFrame}(x)

\S4method{lengths}{MsBackendDataFrame}(x, use.names = FALSE)

\S4method{msLevel}{MsBackendDataFrame}(object, ...)

\S4method{mz}{MsBackendDataFrame}(object)

\S4method{mz}{MsBackendDataFrame}(object) <- value

\S4method{polarity}{MsBackendDataFrame}(object)

\S4method{polarity}{MsBackendDataFrame}(object) <- value

\S4method{precScanNum}{MsBackendDataFrame}(object)

\S4method{precursorCharge}{MsBackendDataFrame}(object)

\S4method{precursorIntensity}{MsBackendDataFrame}(object)

\S4method{precursorMz}{MsBackendDataFrame}(object)

\S4method{replaceList}{MsBackendDataFrame}(object) <- value

\S4method{rtime}{MsBackendDataFrame}(object)

\S4method{rtime}{MsBackendDataFrame}(object) <- value

\S4method{scanIndex}{MsBackendDataFrame}(object)

\S4method{selectSpectraVariables}{MsBackendDataFrame}(object, spectraVariables = spectraVariables(object))

\S4method{smoothed}{MsBackendDataFrame}(object)

\S4method{smoothed}{MsBackendDataFrame}(object) <- value

\S4method{spectraData}{MsBackendDataFrame}(object, columns = spectraVariables(object))

\S4method{spectraData}{MsBackendDataFrame}(object) <- value

\S4method{spectraNames}{MsBackendDataFrame}(object)

\S4method{spectraNames}{MsBackendDataFrame}(object) <- value

\S4method{spectraVariables}{MsBackendDataFrame}(object)

\S4method{tic}{MsBackendDataFrame}(object, initial = TRUE)

\S4method{$}{MsBackendDataFrame}(x, name)

\S4method{$}{MsBackendDataFrame}(x, name) <- value

\S4method{[}{MsBackendDataFrame}(x, i, j, ..., drop = FALSE)

\S4method{split}{MsBackendDataFrame,ANY}(x, f, drop = FALSE, ...)

\S4method{filterAcquisitionNum}{MsBackendDataFrame}(
  object,
  n = integer(),
  dataStorage = character(),
  dataOrigin = character()
)

\S4method{filterDataOrigin}{MsBackendDataFrame}(object, dataOrigin = character())

\S4method{filterDataStorage}{MsBackendDataFrame}(object, dataStorage = character())

\S4method{filterEmptySpectra}{MsBackendDataFrame}(object)

\S4method{filterIsolationWindow}{MsBackendDataFrame}(object, mz = numeric(), ...)

\S4method{filterMsLevel}{MsBackendDataFrame}(object, msLevel = integer())

\S4method{filterPolarity}{MsBackendDataFrame}(object, polarity = integer())

\S4method{filterPrecursorMz}{MsBackendDataFrame}(object, mz = numeric())

\S4method{filterPrecursorScan}{MsBackendDataFrame}(object, acquisitionNum = integer())

\S4method{filterRt}{MsBackendDataFrame}(object, rt = numeric(), msLevel. = unique(msLevel(object)))

\S4method{backendInitialize}{MsBackendHdf5Peaks}(
  object,
  files = character(),
  spectraData = DataFrame(),
  hdf5path = character(),
  ...,
  BPPARAM = bpparam()
)

\S4method{show}{MsBackendHdf5Peaks}(object)

\S4method{as.list}{MsBackendHdf5Peaks}(x)

\S4method{intensity}{MsBackendHdf5Peaks}(object)

\S4method{intensity}{MsBackendHdf5Peaks}(object) <- value

\S4method{ionCount}{MsBackendHdf5Peaks}(object)

\S4method{isCentroided}{MsBackendHdf5Peaks}(object, ...)

\S4method{isEmpty}{MsBackendHdf5Peaks}(x)

\S4method{lengths}{MsBackendHdf5Peaks}(x, use.names = FALSE)

\S4method{mz}{MsBackendHdf5Peaks}(object)

\S4method{mz}{MsBackendHdf5Peaks}(object) <- value

\S4method{replaceList}{MsBackendHdf5Peaks}(object) <- value

\S4method{spectraData}{MsBackendHdf5Peaks}(object, columns = spectraVariables(object))

\S4method{spectraData}{MsBackendHdf5Peaks}(object) <- value

\S4method{$}{MsBackendHdf5Peaks}(x, name) <- value

\S4method{[}{MsBackendHdf5Peaks}(x, i, j, ..., drop = FALSE)

\S4method{backendMerge}{MsBackendHdf5Peaks}(object, ...)

\S4method{backendInitialize}{MsBackendMzR}(object, files, ..., BPPARAM = bpparam())

\S4method{show}{MsBackendMzR}(object)

\S4method{as.list}{MsBackendMzR}(x)

\S4method{intensity}{MsBackendMzR}(object)

\S4method{intensity}{MsBackendMzR}(object) <- value

\S4method{ionCount}{MsBackendMzR}(object)

\S4method{isCentroided}{MsBackendMzR}(object, ...)

\S4method{isEmpty}{MsBackendMzR}(x)

\S4method{lengths}{MsBackendMzR}(x, use.names = FALSE)

\S4method{mz}{MsBackendMzR}(object)

\S4method{mz}{MsBackendMzR}(object) <- value

\S4method{spectraData}{MsBackendMzR}(object, columns = spectraVariables(object))

\S4method{spectraData}{MsBackendMzR}(object) <- value

\S4method{spectraNames}{MsBackendMzR}(object)

\S4method{spectraNames}{MsBackendMzR}(object) <- value

\S4method{spectraVariables}{MsBackendMzR}(object)

\S4method{$}{MsBackendMzR}(x, name) <- value

\S4method{show}{ProcessingStep}(object)

\S4method{show}{Spectra}(object)
}
\description{
For S4 methods that require a documentation entry but only clutter the index.
}
\keyword{internal}
