% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_aliases.R, R/Backend-class.R,
%   R/BackendMzR-class.R, R/Spectra-class.R
\docType{methods}
\name{hidden_aliases}
\alias{hidden_aliases}
\alias{show,Backend-method}
\alias{fileNames,Backend-method}
\alias{backendReadSpectra,BackendMzR-method}
\alias{backendWriteSpectra,BackendMzR-method}
\alias{show,Spectra-method}
\alias{setBackend,Spectra,Backend-method}
\alias{setBackend,Spectra,BackendMzR-method}
\title{Internal page for hidden aliases}
\usage{
\S4method{show}{Backend}(object)

\S4method{fileNames}{Backend}(object, ...)

\S4method{backendReadSpectra}{BackendMzR}(object, spectraData,
  processingQueue = list(), ...)

\S4method{backendWriteSpectra}{BackendMzR}(object, spectra, spectraData)

\S4method{show}{Spectra}(object)

\S4method{setBackend}{Spectra,Backend}(object, backend, ...,
  BPPARAM = bpparam())

\S4method{setBackend}{Spectra,BackendMzR}(object, backend, ...,
  BPPARAM = bpparam())
}
\arguments{
\item{object}{Object to display.}

\item{object}{Object to display.}
}
\description{
For S4 methods that require a documentation entry but only clutter the index.
}
\keyword{internal}
