% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_aliases.R, R/AllGenerics.R,
%   R/MsBackendMemory.R, R/MsBackendMzR.R, R/ProcessingStep.R, R/Spectra.R
\docType{methods}
\name{hidden_aliases}
\alias{hidden_aliases}
\alias{acquisitionNum}
\alias{backendInitialize}
\alias{centroided}
\alias{centroided<-}
\alias{collisionEnergy}
\alias{collisionEnergy<-}
\alias{fileNames}
\alias{fromFile}
\alias{ionCount}
\alias{isReadOnly}
\alias{msLevel}
\alias{peaksCount}
\alias{polarity}
\alias{polarity<-}
\alias{precScanNum}
\alias{precursorCharge}
\alias{precursorIntensity}
\alias{precursorMz}
\alias{rtime<-}
\alias{scanIndex}
\alias{smoothed}
\alias{smoothed<-}
\alias{spectraData}
\alias{spectraData<-}
\alias{spectraNames}
\alias{spectraNames<-}
\alias{spectraVariables}
\alias{backendInitialize,MsBackendMemory-method}
\alias{acquisitionNum,MsBackendMemory-method}
\alias{centroided,MsBackendMemory-method}
\alias{centroided<-,MsBackendMemory-method}
\alias{collisionEnergy,MsBackendMemory-method}
\alias{collisionEnergy<-,MsBackendMemory-method}
\alias{fromFile,MsBackendMemory-method}
\alias{intensity,MsBackendMemory-method}
\alias{ionCount,MsBackendMemory-method}
\alias{isCentroided,MsBackendMemory-method}
\alias{isEmpty,MsBackendMemory-method}
\alias{length,MsBackendMemory-method}
\alias{msLevel,MsBackendMemory-method}
\alias{mz,MsBackendMemory-method}
\alias{peaks,MsBackendMemory-method}
\alias{peaksCount,MsBackendMemory-method}
\alias{polarity,MsBackendMemory-method}
\alias{polarity<-,MsBackendMemory-method}
\alias{precScanNum,MsBackendMemory-method}
\alias{precursorCharge,MsBackendMemory-method}
\alias{precursorIntensity,MsBackendMemory-method}
\alias{precursorMz,MsBackendMemory-method}
\alias{rtime,MsBackendMemory-method}
\alias{rtime<-,MsBackendMemory-method}
\alias{scanIndex,MsBackendMemory-method}
\alias{smoothed,MsBackendMemory-method}
\alias{smoothed<-,MsBackendMemory-method}
\alias{spectraData,MsBackendMemory-method}
\alias{spectraData<-,MsBackendMemory-method}
\alias{spectraNames,MsBackendMemory-method}
\alias{spectraNames<-,MsBackendMemory-method}
\alias{spectraVariables,MsBackendMemory-method}
\alias{tic,MsBackendMemory-method}
\alias{backendInitialize,MsBackendMzR-method}
\alias{ProcessingStep-class}
\alias{show,ProcessingStep-method}
\alias{show,Spectra-method}
\title{Internal page for hidden aliases}
\usage{
acquisitionNum(object, ...)

backendInitialize(object, files, spectraData, ...)

centroided(object, ...)

centroided(object) <- value

collisionEnergy(object, ...)

collisionEnergy(object) <- value

fileNames(object, ...)

fromFile(object, ...)

ionCount(object, ...)

isReadOnly(object, ...)

msLevel(object, ...)

peaksCount(object, ...)

polarity(object, ...)

polarity(object) <- value

precScanNum(object, ...)

precursorCharge(object, ...)

precursorIntensity(object, ...)

precursorMz(object, ...)

rtime(object) <- value

scanIndex(object, ...)

smoothed(object, ...)

smoothed(object) <- value

spectraData(object, ...)

spectraData(object) <- value

spectraNames(object, ...)

spectraNames(object) <- value

spectraVariables(object, ...)

\S4method{backendInitialize}{MsBackendMemory}(object, files, spectraData,
  ...)

\S4method{acquisitionNum}{MsBackendMemory}(object)

\S4method{centroided}{MsBackendMemory}(object)

\S4method{centroided}{MsBackendMemory}(object) <- value

\S4method{collisionEnergy}{MsBackendMemory}(object)

\S4method{collisionEnergy}{MsBackendMemory}(object) <- value

\S4method{fromFile}{MsBackendMemory}(object)

\S4method{intensity}{MsBackendMemory}(object)

\S4method{ionCount}{MsBackendMemory}(object)

\S4method{isCentroided}{MsBackendMemory}(object, ...)

\S4method{isEmpty}{MsBackendMemory}(x)

\S4method{length}{MsBackendMemory}(x)

\S4method{msLevel}{MsBackendMemory}(object, ...)

\S4method{mz}{MsBackendMemory}(object)

\S4method{peaks}{MsBackendMemory}(object)

\S4method{peaksCount}{MsBackendMemory}(object)

\S4method{polarity}{MsBackendMemory}(object)

\S4method{polarity}{MsBackendMemory}(object) <- value

\S4method{precScanNum}{MsBackendMemory}(object)

\S4method{precursorCharge}{MsBackendMemory}(object)

\S4method{precursorIntensity}{MsBackendMemory}(object)

\S4method{precursorMz}{MsBackendMemory}(object)

\S4method{rtime}{MsBackendMemory}(object)

\S4method{rtime}{MsBackendMemory}(object) <- value

\S4method{scanIndex}{MsBackendMemory}(object)

\S4method{smoothed}{MsBackendMemory}(object)

\S4method{smoothed}{MsBackendMemory}(object) <- value

\S4method{spectraData}{MsBackendMemory}(object, columns)

\S4method{spectraData}{MsBackendMemory}(object) <- value

\S4method{spectraNames}{MsBackendMemory}(object)

\S4method{spectraNames}{MsBackendMemory}(object) <- value

\S4method{spectraVariables}{MsBackendMemory}(object)

\S4method{tic}{MsBackendMemory}(object, initial = TRUE)

\S4method{backendInitialize}{MsBackendMzR}(object, files, spectraData, ...,
  BPPARAM = bpparam())

\S4method{show}{ProcessingStep}(object)

\S4method{show}{Spectra}(object)
}
\description{
For S4 methods that require a documentation entry but only clutter the index.
}
\keyword{internal}
