% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_aliases.R, R/AllGenerics.R,
%   R/MsBackendMemory.R, R/MsBackendMzR.R, R/ProcessingStep.R, R/Spectra.R
\docType{methods}
\name{hidden_aliases}
\alias{hidden_aliases}
\alias{acquisitionNum}
\alias{backendInitialize}
\alias{centroided}
\alias{centroided<-}
\alias{collisionEnergy}
\alias{collisionEnergy<-}
\alias{fileNames}
\alias{fromFile}
\alias{ionCount}
\alias{isReadOnly}
\alias{msLevel}
\alias{peaksCount}
\alias{polarity}
\alias{polarity<-}
\alias{precScanNum}
\alias{precursorCharge}
\alias{precursorIntensity}
\alias{precursorMz}
\alias{rtime<-}
\alias{scanIndex}
\alias{smoothed}
\alias{smoothed<-}
\alias{spectraData}
\alias{spectraData<-}
\alias{spectraNames}
\alias{spectraNames<-}
\alias{spectraVariables}
\alias{backendInitialize,MsBackendMemory-method}
\alias{length,MsBackendMemory-method}
\alias{msLevel,MsBackendMemory-method}
\alias{backendInitialize,MsBackendMzR-method}
\alias{ProcessingStep-class}
\alias{show,ProcessingStep-method}
\alias{show,Spectra-method}
\title{Internal page for hidden aliases}
\usage{
acquisitionNum(object, ...)

backendInitialize(object, files, spectraData, ...)

centroided(object, ...)

centroided(object) <- value

collisionEnergy(object, ...)

collisionEnergy(object) <- value

fileNames(object, ...)

fromFile(object, ...)

ionCount(object, ...)

isReadOnly(object, ...)

msLevel(object, ...)

peaksCount(object, ...)

polarity(object, ...)

polarity(object) <- value

precScanNum(object, ...)

precursorCharge(object, ...)

precursorIntensity(object, ...)

precursorMz(object, ...)

rtime(object) <- value

scanIndex(object, ...)

smoothed(object, ...)

smoothed(object) <- value

spectraData(object, ...)

spectraData(object) <- value

spectraNames(object, ...)

spectraNames(object) <- value

spectraVariables(object, ...)

\S4method{backendInitialize}{MsBackendMemory}(object, files, spectraData,
  ...)

\S4method{length}{MsBackendMemory}(x)

\S4method{msLevel}{MsBackendMemory}(object, ...)

\S4method{backendInitialize}{MsBackendMzR}(object, files, spectraData, ...,
  BPPARAM = bpparam())

\S4method{show}{ProcessingStep}(object)

\S4method{show}{Spectra}(object)
}
\description{
For S4 methods that require a documentation entry but only clutter the index.
}
\keyword{internal}
