% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_aliases.R, R/AllGenerics.R,
%   R/MsBackendDataFrame.R, R/MsBackendMzR.R, R/MsBackendRleDataFrame.R,
%   R/ProcessingStep.R, R/Spectra.R
\docType{methods}
\name{hidden_aliases}
\alias{hidden_aliases}
\alias{acquisitionNum}
\alias{backendInitialize}
\alias{centroided}
\alias{centroided<-}
\alias{collisionEnergy}
\alias{collisionEnergy<-}
\alias{fileNames}
\alias{fromFile}
\alias{ionCount}
\alias{isReadOnly}
\alias{msLevel}
\alias{peaksCount}
\alias{polarity}
\alias{polarity<-}
\alias{precScanNum}
\alias{precursorCharge}
\alias{precursorIntensity}
\alias{precursorMz}
\alias{rtime<-}
\alias{scanIndex}
\alias{smoothed}
\alias{smoothed<-}
\alias{spectraData}
\alias{spectraData<-}
\alias{spectraNames}
\alias{spectraNames<-}
\alias{spectraVariables}
\alias{show,MsBackendDataFrame-method}
\alias{backendInitialize,MsBackendDataFrame-method}
\alias{acquisitionNum,MsBackendDataFrame-method}
\alias{centroided,MsBackendDataFrame-method}
\alias{centroided<-,MsBackendDataFrame-method}
\alias{collisionEnergy,MsBackendDataFrame-method}
\alias{collisionEnergy<-,MsBackendDataFrame-method}
\alias{fromFile,MsBackendDataFrame-method}
\alias{intensity,MsBackendDataFrame-method}
\alias{ionCount,MsBackendDataFrame-method}
\alias{isCentroided,MsBackendDataFrame-method}
\alias{isEmpty,MsBackendDataFrame-method}
\alias{length,MsBackendDataFrame-method}
\alias{msLevel,MsBackendDataFrame-method}
\alias{mz,MsBackendDataFrame-method}
\alias{peaks,MsBackendDataFrame-method}
\alias{peaksCount,MsBackendDataFrame-method}
\alias{polarity,MsBackendDataFrame-method}
\alias{polarity<-,MsBackendDataFrame-method}
\alias{precScanNum,MsBackendDataFrame-method}
\alias{precursorCharge,MsBackendDataFrame-method}
\alias{precursorIntensity,MsBackendDataFrame-method}
\alias{precursorMz,MsBackendDataFrame-method}
\alias{rtime,MsBackendDataFrame-method}
\alias{rtime<-,MsBackendDataFrame-method}
\alias{scanIndex,MsBackendDataFrame-method}
\alias{smoothed,MsBackendDataFrame-method}
\alias{smoothed<-,MsBackendDataFrame-method}
\alias{spectraData,MsBackendDataFrame-method}
\alias{spectraData<-,MsBackendDataFrame-method}
\alias{spectraNames,MsBackendDataFrame-method}
\alias{spectraNames<-,MsBackendDataFrame-method}
\alias{spectraVariables,MsBackendDataFrame-method}
\alias{tic,MsBackendDataFrame-method}
\alias{[,MsBackendDataFrame-method}
\alias{backendInitialize,MsBackendMzR-method}
\alias{show,MsBackendRleDataFrame-method}
\alias{backendInitialize,MsBackendRleDataFrame-method}
\alias{acquisitionNum,MsBackendRleDataFrame-method}
\alias{centroided,MsBackendRleDataFrame-method}
\alias{centroided<-,MsBackendRleDataFrame-method}
\alias{collisionEnergy,MsBackendRleDataFrame-method}
\alias{collisionEnergy<-,MsBackendRleDataFrame-method}
\alias{fromFile,MsBackendRleDataFrame-method}
\alias{intensity,MsBackendRleDataFrame-method}
\alias{ionCount,MsBackendRleDataFrame-method}
\alias{isCentroided,MsBackendRleDataFrame-method}
\alias{isEmpty,MsBackendRleDataFrame-method}
\alias{length,MsBackendRleDataFrame-method}
\alias{msLevel,MsBackendRleDataFrame-method}
\alias{mz,MsBackendRleDataFrame-method}
\alias{peaks,MsBackendRleDataFrame-method}
\alias{peaksCount,MsBackendRleDataFrame-method}
\alias{polarity,MsBackendRleDataFrame-method}
\alias{polarity<-,MsBackendRleDataFrame-method}
\alias{precScanNum,MsBackendRleDataFrame-method}
\alias{precursorCharge,MsBackendRleDataFrame-method}
\alias{precursorIntensity,MsBackendRleDataFrame-method}
\alias{precursorMz,MsBackendRleDataFrame-method}
\alias{rtime,MsBackendRleDataFrame-method}
\alias{rtime<-,MsBackendRleDataFrame-method}
\alias{scanIndex,MsBackendRleDataFrame-method}
\alias{smoothed,MsBackendRleDataFrame-method}
\alias{smoothed<-,MsBackendRleDataFrame-method}
\alias{spectraData,MsBackendRleDataFrame-method}
\alias{spectraData<-,MsBackendRleDataFrame-method}
\alias{spectraNames,MsBackendRleDataFrame-method}
\alias{spectraNames<-,MsBackendRleDataFrame-method}
\alias{spectraVariables,MsBackendRleDataFrame-method}
\alias{tic,MsBackendRleDataFrame-method}
\alias{[,MsBackendRleDataFrame-method}
\alias{ProcessingStep-class}
\alias{show,ProcessingStep-method}
\alias{show,Spectra-method}
\title{Internal page for hidden aliases}
\usage{
acquisitionNum(object, ...)

backendInitialize(object, files, spectraData, ...)

centroided(object, ...)

centroided(object) <- value

collisionEnergy(object, ...)

collisionEnergy(object) <- value

fileNames(object, ...)

fromFile(object, ...)

ionCount(object, ...)

isReadOnly(object, ...)

msLevel(object, ...)

peaksCount(object, ...)

polarity(object, ...)

polarity(object) <- value

precScanNum(object, ...)

precursorCharge(object, ...)

precursorIntensity(object, ...)

precursorMz(object, ...)

rtime(object) <- value

scanIndex(object, ...)

smoothed(object, ...)

smoothed(object) <- value

spectraData(object, ...)

spectraData(object) <- value

spectraNames(object, ...)

spectraNames(object) <- value

spectraVariables(object, ...)

\S4method{show}{MsBackendDataFrame}(object)

\S4method{backendInitialize}{MsBackendDataFrame}(object, files,
  spectraData, ...)

\S4method{acquisitionNum}{MsBackendDataFrame}(object)

\S4method{centroided}{MsBackendDataFrame}(object)

\S4method{centroided}{MsBackendDataFrame}(object) <- value

\S4method{collisionEnergy}{MsBackendDataFrame}(object)

\S4method{collisionEnergy}{MsBackendDataFrame}(object) <- value

\S4method{fromFile}{MsBackendDataFrame}(object)

\S4method{intensity}{MsBackendDataFrame}(object)

\S4method{ionCount}{MsBackendDataFrame}(object)

\S4method{isCentroided}{MsBackendDataFrame}(object, ...)

\S4method{isEmpty}{MsBackendDataFrame}(x)

\S4method{length}{MsBackendDataFrame}(x)

\S4method{msLevel}{MsBackendDataFrame}(object, ...)

\S4method{mz}{MsBackendDataFrame}(object)

\S4method{peaks}{MsBackendDataFrame}(object)

\S4method{peaksCount}{MsBackendDataFrame}(object)

\S4method{polarity}{MsBackendDataFrame}(object)

\S4method{polarity}{MsBackendDataFrame}(object) <- value

\S4method{precScanNum}{MsBackendDataFrame}(object)

\S4method{precursorCharge}{MsBackendDataFrame}(object)

\S4method{precursorIntensity}{MsBackendDataFrame}(object)

\S4method{precursorMz}{MsBackendDataFrame}(object)

\S4method{rtime}{MsBackendDataFrame}(object)

\S4method{rtime}{MsBackendDataFrame}(object) <- value

\S4method{scanIndex}{MsBackendDataFrame}(object)

\S4method{smoothed}{MsBackendDataFrame}(object)

\S4method{smoothed}{MsBackendDataFrame}(object) <- value

\S4method{spectraData}{MsBackendDataFrame}(object,
  columns = spectraVariables(object))

\S4method{spectraData}{MsBackendDataFrame}(object) <- value

\S4method{spectraNames}{MsBackendDataFrame}(object)

\S4method{spectraNames}{MsBackendDataFrame}(object) <- value

\S4method{spectraVariables}{MsBackendDataFrame}(object)

\S4method{tic}{MsBackendDataFrame}(object, initial = TRUE)

\S4method{[}{MsBackendDataFrame}(x, i, j, ..., drop = FALSE)

\S4method{backendInitialize}{MsBackendMzR}(object, files, spectraData, ...,
  BPPARAM = bpparam())

\S4method{show}{MsBackendRleDataFrame}(object)

\S4method{backendInitialize}{MsBackendRleDataFrame}(object, files,
  spectraData, ...)

\S4method{acquisitionNum}{MsBackendRleDataFrame}(object)

\S4method{centroided}{MsBackendRleDataFrame}(object)

\S4method{centroided}{MsBackendRleDataFrame}(object) <- value

\S4method{collisionEnergy}{MsBackendRleDataFrame}(object)

\S4method{collisionEnergy}{MsBackendRleDataFrame}(object) <- value

\S4method{fromFile}{MsBackendRleDataFrame}(object)

\S4method{intensity}{MsBackendRleDataFrame}(object)

\S4method{ionCount}{MsBackendRleDataFrame}(object)

\S4method{isCentroided}{MsBackendRleDataFrame}(object, ...)

\S4method{isEmpty}{MsBackendRleDataFrame}(x)

\S4method{length}{MsBackendRleDataFrame}(x)

\S4method{msLevel}{MsBackendRleDataFrame}(object, ...)

\S4method{mz}{MsBackendRleDataFrame}(object)

\S4method{peaks}{MsBackendRleDataFrame}(object)

\S4method{peaksCount}{MsBackendRleDataFrame}(object)

\S4method{polarity}{MsBackendRleDataFrame}(object)

\S4method{polarity}{MsBackendRleDataFrame}(object) <- value

\S4method{precScanNum}{MsBackendRleDataFrame}(object)

\S4method{precursorCharge}{MsBackendRleDataFrame}(object)

\S4method{precursorIntensity}{MsBackendRleDataFrame}(object)

\S4method{precursorMz}{MsBackendRleDataFrame}(object)

\S4method{rtime}{MsBackendRleDataFrame}(object)

\S4method{rtime}{MsBackendRleDataFrame}(object) <- value

\S4method{scanIndex}{MsBackendRleDataFrame}(object)

\S4method{smoothed}{MsBackendRleDataFrame}(object)

\S4method{smoothed}{MsBackendRleDataFrame}(object) <- value

\S4method{spectraData}{MsBackendRleDataFrame}(object,
  columns = spectraVariables(object))

\S4method{spectraData}{MsBackendRleDataFrame}(object) <- value

\S4method{spectraNames}{MsBackendRleDataFrame}(object)

\S4method{spectraNames}{MsBackendRleDataFrame}(object) <- value

\S4method{spectraVariables}{MsBackendRleDataFrame}(object)

\S4method{tic}{MsBackendRleDataFrame}(object, initial = TRUE)

\S4method{[}{MsBackendRleDataFrame}(x, i, j, ..., drop = FALSE)

\S4method{show}{ProcessingStep}(object)

\S4method{show}{Spectra}(object)
}
\description{
For S4 methods that require a documentation entry but only clutter the index.
}
\keyword{internal}
