% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_aliases.R, R/AllGenerics.R,
%   R/MsBackendDataFrame.R, R/MsBackendHdf5Peaks.R, R/MsBackendMzR.R,
%   R/Spectra.R
\docType{methods}
\name{hidden_aliases}
\alias{hidden_aliases}
\alias{[,MsBackendDataFrame-method}
\alias{ppm}
\alias{addProcessing}
\alias{backendInitialize}
\alias{backendMerge}
\alias{bin,numeric-method}
\alias{containsMz}
\alias{containsNeutralLoss}
\alias{dropNaSpectraVariables}
\alias{export}
\alias{filterMzRange}
\alias{filterMzValues}
\alias{filterPrecursorMzValues}
\alias{filterPrecursorMzRange}
\alias{isReadOnly}
\alias{peaksData}
\alias{peaksData<-}
\alias{peaksVariables}
\alias{pickPeaks}
\alias{replaceIntensitiesBelow}
\alias{reset}
\alias{selectSpectraVariables}
\alias{setBackend}
\alias{spectrapply}
\alias{show,MsBackendDataFrame-method}
\alias{backendMerge,MsBackendDataFrame-method}
\alias{acquisitionNum,MsBackendDataFrame-method}
\alias{peaksData,MsBackendDataFrame-method}
\alias{centroided,MsBackendDataFrame-method}
\alias{centroided<-,MsBackendDataFrame-method}
\alias{collisionEnergy,MsBackendDataFrame-method}
\alias{collisionEnergy<-,MsBackendDataFrame-method}
\alias{dataOrigin,MsBackendDataFrame-method}
\alias{dataOrigin<-,MsBackendDataFrame-method}
\alias{dataStorage,MsBackendDataFrame-method}
\alias{dataStorage<-,MsBackendDataFrame-method}
\alias{intensity,MsBackendDataFrame-method}
\alias{intensity<-,MsBackendDataFrame-method}
\alias{ionCount,MsBackendDataFrame-method}
\alias{isEmpty,MsBackendDataFrame-method}
\alias{isolationWindowLowerMz,MsBackendDataFrame-method}
\alias{isolationWindowLowerMz<-,MsBackendDataFrame-method}
\alias{isolationWindowTargetMz,MsBackendDataFrame-method}
\alias{isolationWindowTargetMz<-,MsBackendDataFrame-method}
\alias{isolationWindowUpperMz,MsBackendDataFrame-method}
\alias{isolationWindowUpperMz<-,MsBackendDataFrame-method}
\alias{length,MsBackendDataFrame-method}
\alias{lengths,MsBackendDataFrame-method}
\alias{msLevel,MsBackendDataFrame-method}
\alias{msLevel<-,MsBackendDataFrame-method}
\alias{mz,MsBackendDataFrame-method}
\alias{mz<-,MsBackendDataFrame-method}
\alias{polarity,MsBackendDataFrame-method}
\alias{polarity<-,MsBackendDataFrame-method}
\alias{precScanNum,MsBackendDataFrame-method}
\alias{precursorCharge,MsBackendDataFrame-method}
\alias{precursorIntensity,MsBackendDataFrame-method}
\alias{precursorMz,MsBackendDataFrame-method}
\alias{peaksData<-,MsBackendDataFrame-method}
\alias{rtime,MsBackendDataFrame-method}
\alias{rtime<-,MsBackendDataFrame-method}
\alias{scanIndex,MsBackendDataFrame-method}
\alias{selectSpectraVariables,MsBackendDataFrame-method}
\alias{smoothed,MsBackendDataFrame-method}
\alias{smoothed<-,MsBackendDataFrame-method}
\alias{spectraData,MsBackendDataFrame-method}
\alias{spectraData<-,MsBackendDataFrame-method}
\alias{spectraNames,MsBackendDataFrame-method}
\alias{spectraNames<-,MsBackendDataFrame-method}
\alias{spectraVariables,MsBackendDataFrame-method}
\alias{tic,MsBackendDataFrame-method}
\alias{$,MsBackendDataFrame-method}
\alias{$<-,MsBackendDataFrame-method}
\alias{split,MsBackendDataFrame,ANY-method}
\alias{filterAcquisitionNum,MsBackendDataFrame-method}
\alias{backendInitialize,MsBackendHdf5Peaks-method}
\alias{show,MsBackendHdf5Peaks-method}
\alias{peaksData,MsBackendHdf5Peaks-method}
\alias{intensity,MsBackendHdf5Peaks-method}
\alias{intensity<-,MsBackendHdf5Peaks-method}
\alias{ionCount,MsBackendHdf5Peaks-method}
\alias{isCentroided,MsBackendHdf5Peaks-method}
\alias{isEmpty,MsBackendHdf5Peaks-method}
\alias{lengths,MsBackendHdf5Peaks-method}
\alias{mz,MsBackendHdf5Peaks-method}
\alias{mz<-,MsBackendHdf5Peaks-method}
\alias{peaksData<-,MsBackendHdf5Peaks-method}
\alias{spectraData,MsBackendHdf5Peaks-method}
\alias{spectraData<-,MsBackendHdf5Peaks-method}
\alias{$<-,MsBackendHdf5Peaks-method}
\alias{[,MsBackendHdf5Peaks-method}
\alias{backendMerge,MsBackendHdf5Peaks-method}
\alias{backendInitialize,MsBackendMzR-method}
\alias{show,MsBackendMzR-method}
\alias{peaksData,MsBackendMzR-method}
\alias{intensity,MsBackendMzR-method}
\alias{intensity<-,MsBackendMzR-method}
\alias{ionCount,MsBackendMzR-method}
\alias{isCentroided,MsBackendMzR-method}
\alias{isEmpty,MsBackendMzR-method}
\alias{lengths,MsBackendMzR-method}
\alias{mz,MsBackendMzR-method}
\alias{mz<-,MsBackendMzR-method}
\alias{spectraData,MsBackendMzR-method}
\alias{spectraData<-,MsBackendMzR-method}
\alias{spectraNames,MsBackendMzR-method}
\alias{spectraNames<-,MsBackendMzR-method}
\alias{spectraVariables,MsBackendMzR-method}
\alias{$<-,MsBackendMzR-method}
\alias{export,MsBackendMzR-method}
\alias{show,Spectra-method}
\title{Internal page for hidden aliases}
\usage{
addProcessing(object, ...)

backendInitialize(object, ...)

backendMerge(object, ...)

\S4method{bin}{numeric}(
  x,
  y,
  size = 1,
  breaks = seq(floor(min(y)), ceiling(max(y)), by = size),
  FUN = max
)

containsMz(object, ...)

containsNeutralLoss(object, ...)

dropNaSpectraVariables(object, ...)

export(object, ...)

filterMzRange(object, ...)

filterMzValues(object, ...)

filterPrecursorMzValues(object, ...)

filterPrecursorMzRange(object, ...)

isReadOnly(object, ...)

peaksData(object, ...)

peaksData(object) <- value

peaksVariables(object, ...)

pickPeaks(object, ...)

replaceIntensitiesBelow(object, threshold = min, ...)

reset(object, ...)

selectSpectraVariables(object, ...)

setBackend(object, backend, ...)

spectrapply(object, ...)

\S4method{show}{MsBackendDataFrame}(object)

\S4method{backendMerge}{MsBackendDataFrame}(object, ...)

\S4method{acquisitionNum}{MsBackendDataFrame}(object)

\S4method{peaksData}{MsBackendDataFrame}(object, columns = peaksVariables(object), ...)

\S4method{centroided}{MsBackendDataFrame}(object)

\S4method{centroided}{MsBackendDataFrame}(object) <- value

\S4method{collisionEnergy}{MsBackendDataFrame}(object)

\S4method{collisionEnergy}{MsBackendDataFrame}(object) <- value

\S4method{dataOrigin}{MsBackendDataFrame}(object)

\S4method{dataOrigin}{MsBackendDataFrame}(object) <- value

\S4method{dataStorage}{MsBackendDataFrame}(object)

\S4method{dataStorage}{MsBackendDataFrame}(object) <- value

\S4method{intensity}{MsBackendDataFrame}(object)

\S4method{intensity}{MsBackendDataFrame}(object) <- value

\S4method{ionCount}{MsBackendDataFrame}(object)

\S4method{isEmpty}{MsBackendDataFrame}(x)

\S4method{isolationWindowLowerMz}{MsBackendDataFrame}(object)

\S4method{isolationWindowLowerMz}{MsBackendDataFrame}(object) <- value

\S4method{isolationWindowTargetMz}{MsBackendDataFrame}(object)

\S4method{isolationWindowTargetMz}{MsBackendDataFrame}(object) <- value

\S4method{isolationWindowUpperMz}{MsBackendDataFrame}(object)

\S4method{isolationWindowUpperMz}{MsBackendDataFrame}(object) <- value

\S4method{length}{MsBackendDataFrame}(x)

\S4method{lengths}{MsBackendDataFrame}(x, use.names = FALSE)

\S4method{msLevel}{MsBackendDataFrame}(object, ...)

\S4method{msLevel}{MsBackendDataFrame}(object) <- value

\S4method{mz}{MsBackendDataFrame}(object)

\S4method{mz}{MsBackendDataFrame}(object) <- value

\S4method{polarity}{MsBackendDataFrame}(object)

\S4method{polarity}{MsBackendDataFrame}(object) <- value

\S4method{precScanNum}{MsBackendDataFrame}(object)

\S4method{precursorCharge}{MsBackendDataFrame}(object)

\S4method{precursorIntensity}{MsBackendDataFrame}(object)

\S4method{precursorMz}{MsBackendDataFrame}(object)

\S4method{peaksData}{MsBackendDataFrame}(object) <- value

\S4method{rtime}{MsBackendDataFrame}(object)

\S4method{rtime}{MsBackendDataFrame}(object) <- value

\S4method{scanIndex}{MsBackendDataFrame}(object)

\S4method{selectSpectraVariables}{MsBackendDataFrame}(object, spectraVariables = spectraVariables(object))

\S4method{smoothed}{MsBackendDataFrame}(object)

\S4method{smoothed}{MsBackendDataFrame}(object) <- value

\S4method{spectraData}{MsBackendDataFrame}(object, columns = spectraVariables(object))

\S4method{spectraData}{MsBackendDataFrame}(object) <- value

\S4method{spectraNames}{MsBackendDataFrame}(object)

\S4method{spectraNames}{MsBackendDataFrame}(object) <- value

\S4method{spectraVariables}{MsBackendDataFrame}(object)

\S4method{tic}{MsBackendDataFrame}(object, initial = TRUE)

\S4method{$}{MsBackendDataFrame}(x, name)

\S4method{$}{MsBackendDataFrame}(x, name) <- value

\S4method{[}{MsBackendDataFrame}(x, i, j, ..., drop = FALSE)

\S4method{split}{MsBackendDataFrame,ANY}(x, f, drop = FALSE, ...)

\S4method{filterAcquisitionNum}{MsBackendDataFrame}(
  object,
  n = integer(),
  dataStorage = character(),
  dataOrigin = character()
)

\S4method{backendInitialize}{MsBackendHdf5Peaks}(
  object,
  files = character(),
  data = DataFrame(),
  hdf5path = character(),
  ...,
  BPPARAM = bpparam()
)

\S4method{show}{MsBackendHdf5Peaks}(object)

\S4method{peaksData}{MsBackendHdf5Peaks}(object, ...)

\S4method{intensity}{MsBackendHdf5Peaks}(object)

\S4method{intensity}{MsBackendHdf5Peaks}(object) <- value

\S4method{ionCount}{MsBackendHdf5Peaks}(object)

\S4method{isCentroided}{MsBackendHdf5Peaks}(object, ...)

\S4method{isEmpty}{MsBackendHdf5Peaks}(x)

\S4method{lengths}{MsBackendHdf5Peaks}(x, use.names = FALSE)

\S4method{mz}{MsBackendHdf5Peaks}(object)

\S4method{mz}{MsBackendHdf5Peaks}(object) <- value

\S4method{peaksData}{MsBackendHdf5Peaks}(object) <- value

\S4method{spectraData}{MsBackendHdf5Peaks}(object, columns = spectraVariables(object))

\S4method{spectraData}{MsBackendHdf5Peaks}(object) <- value

\S4method{$}{MsBackendHdf5Peaks}(x, name) <- value

\S4method{[}{MsBackendHdf5Peaks}(x, i, j, ..., drop = FALSE)

\S4method{backendMerge}{MsBackendHdf5Peaks}(object, ...)

\S4method{backendInitialize}{MsBackendMzR}(object, files, ..., BPPARAM = bpparam())

\S4method{show}{MsBackendMzR}(object)

\S4method{peaksData}{MsBackendMzR}(object, ...)

\S4method{intensity}{MsBackendMzR}(object)

\S4method{intensity}{MsBackendMzR}(object) <- value

\S4method{ionCount}{MsBackendMzR}(object)

\S4method{isCentroided}{MsBackendMzR}(object, ...)

\S4method{isEmpty}{MsBackendMzR}(x)

\S4method{lengths}{MsBackendMzR}(x, use.names = FALSE)

\S4method{mz}{MsBackendMzR}(object)

\S4method{mz}{MsBackendMzR}(object) <- value

\S4method{spectraData}{MsBackendMzR}(object, columns = spectraVariables(object))

\S4method{spectraData}{MsBackendMzR}(object) <- value

\S4method{spectraNames}{MsBackendMzR}(object)

\S4method{spectraNames}{MsBackendMzR}(object) <- value

\S4method{spectraVariables}{MsBackendMzR}(object)

\S4method{$}{MsBackendMzR}(x, name) <- value

\S4method{export}{MsBackendMzR}(
  object,
  x,
  file = tempfile(),
  format = c("mzML", "mzXML"),
  copy = FALSE,
  BPPARAM = bpparam()
)

\S4method{show}{Spectra}(object)
}
\value{
Not applicable
}
\description{
For S4 methods that require a documentation entry but only clutter the index.
}
\keyword{internal}
