% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_aliases.R, R/AllGenerics.R,
%   R/MsBackendDataFrame.R, R/MsBackendMzR.R, R/ProcessingStep.R, R/Spectra.R
\docType{methods}
\name{hidden_aliases}
\alias{hidden_aliases}
\alias{[,MsBackendDataFrame-method}
\alias{backendInitialize}
\alias{clean}
\alias{fileNames}
\alias{filterAcquisitionNum}
\alias{filterEmptySpectra}
\alias{filterFile}
\alias{filterMsLevel}
\alias{filterPolarity}
\alias{filterPrecursorScan}
\alias{filterRt}
\alias{fromFile}
\alias{ionCount}
\alias{isReadOnly}
\alias{peaksCount}
\alias{removePeaks}
\alias{selectSpectraVariables}
\alias{show,MsBackendDataFrame-method}
\alias{backendInitialize,MsBackendDataFrame-method}
\alias{acquisitionNum,MsBackendDataFrame-method}
\alias{centroided,MsBackendDataFrame-method}
\alias{centroided<-,MsBackendDataFrame-method}
\alias{collisionEnergy,MsBackendDataFrame-method}
\alias{collisionEnergy<-,MsBackendDataFrame-method}
\alias{fromFile,MsBackendDataFrame-method}
\alias{intensity,MsBackendDataFrame-method}
\alias{ionCount,MsBackendDataFrame-method}
\alias{isCentroided,MsBackendDataFrame-method}
\alias{isEmpty,MsBackendDataFrame-method}
\alias{length,MsBackendDataFrame-method}
\alias{msLevel,MsBackendDataFrame-method}
\alias{mz,MsBackendDataFrame-method}
\alias{peaks,MsBackendDataFrame-method}
\alias{peaksCount,MsBackendDataFrame-method}
\alias{polarity,MsBackendDataFrame-method}
\alias{polarity<-,MsBackendDataFrame-method}
\alias{precScanNum,MsBackendDataFrame-method}
\alias{precursorCharge,MsBackendDataFrame-method}
\alias{precursorIntensity,MsBackendDataFrame-method}
\alias{precursorMz,MsBackendDataFrame-method}
\alias{rtime,MsBackendDataFrame-method}
\alias{rtime<-,MsBackendDataFrame-method}
\alias{scanIndex,MsBackendDataFrame-method}
\alias{selectSpectraVariables,MsBackendDataFrame-method}
\alias{smoothed,MsBackendDataFrame-method}
\alias{smoothed<-,MsBackendDataFrame-method}
\alias{spectraData,MsBackendDataFrame-method}
\alias{spectraData<-,MsBackendDataFrame-method}
\alias{spectraNames,MsBackendDataFrame-method}
\alias{spectraNames<-,MsBackendDataFrame-method}
\alias{spectraVariables,MsBackendDataFrame-method}
\alias{tic,MsBackendDataFrame-method}
\alias{$,MsBackendDataFrame-method}
\alias{$<-,MsBackendDataFrame-method}
\alias{[,MsBackendDataFrame,ANY-method}
\alias{filterAcquisitionNum,MsBackendDataFrame-method}
\alias{filterEmptySpectra,MsBackendDataFrame-method}
\alias{filterFile,MsBackendDataFrame-method}
\alias{backendInitialize,MsBackendMzR-method}
\alias{show,MsBackendMzR-method}
\alias{acquisitionNum,MsBackendMzR-method}
\alias{centroided,MsBackendMzR-method}
\alias{centroided<-,MsBackendMzR-method}
\alias{collisionEnergy,MsBackendMzR-method}
\alias{collisionEnergy<-,MsBackendMzR-method}
\alias{fromFile,MsBackendMzR-method}
\alias{intensity,MsBackendMzR-method}
\alias{ionCount,MsBackendMzR-method}
\alias{isCentroided,MsBackendMzR-method}
\alias{isEmpty,MsBackendMzR-method}
\alias{msLevel,MsBackendMzR-method}
\alias{mz,MsBackendMzR-method}
\alias{peaks,MsBackendMzR-method}
\alias{peaksCount,MsBackendMzR-method}
\alias{polarity,MsBackendMzR-method}
\alias{polarity<-,MsBackendMzR-method}
\alias{precScanNum,MsBackendMzR-method}
\alias{precursorCharge,MsBackendMzR-method}
\alias{precursorIntensity,MsBackendMzR-method}
\alias{precursorMz,MsBackendMzR-method}
\alias{rtime,MsBackendMzR-method}
\alias{rtime<-,MsBackendMzR-method}
\alias{scanIndex,MsBackendMzR-method}
\alias{smoothed,MsBackendMzR-method}
\alias{smoothed<-,MsBackendMzR-method}
\alias{spectraData,MsBackendMzR-method}
\alias{spectraData<-,MsBackendMzR-method}
\alias{spectraNames,MsBackendMzR-method}
\alias{spectraNames<-,MsBackendMzR-method}
\alias{spectraVariables,MsBackendMzR-method}
\alias{tic,MsBackendMzR-method}
\alias{$<-,MsBackendMzR-method}
\alias{ProcessingStep-class}
\alias{show,ProcessingStep-method}
\alias{show,Spectra-method}
\title{Internal page for hidden aliases}
\usage{
backendInitialize(object, files, spectraData, ...)

clean(object, ...)

fileNames(object, ...)

filterAcquisitionNum(object, ...)

filterEmptySpectra(object, ...)

filterFile(object, ...)

filterMsLevel(object, ...)

filterPolarity(object, ...)

filterPrecursorScan(object, ...)

filterRt(object, ...)

fromFile(object, ...)

ionCount(object, ...)

isReadOnly(object, ...)

peaksCount(object, ...)

removePeaks(object, t = "min", ...)

selectSpectraVariables(object, ...)

\S4method{show}{MsBackendDataFrame}(object)

\S4method{backendInitialize}{MsBackendDataFrame}(object, files,
  spectraData, ...)

\S4method{acquisitionNum}{MsBackendDataFrame}(object)

\S4method{centroided}{MsBackendDataFrame}(object)

\S4method{centroided}{MsBackendDataFrame}(object) <- value

\S4method{collisionEnergy}{MsBackendDataFrame}(object)

\S4method{collisionEnergy}{MsBackendDataFrame}(object) <- value

\S4method{fromFile}{MsBackendDataFrame}(object)

\S4method{intensity}{MsBackendDataFrame}(object)

\S4method{ionCount}{MsBackendDataFrame}(object)

\S4method{isCentroided}{MsBackendDataFrame}(object, ...)

\S4method{isEmpty}{MsBackendDataFrame}(x)

\S4method{length}{MsBackendDataFrame}(x)

\S4method{msLevel}{MsBackendDataFrame}(object, ...)

\S4method{mz}{MsBackendDataFrame}(object)

\S4method{peaks}{MsBackendDataFrame}(object)

\S4method{peaksCount}{MsBackendDataFrame}(object)

\S4method{polarity}{MsBackendDataFrame}(object)

\S4method{polarity}{MsBackendDataFrame}(object) <- value

\S4method{precScanNum}{MsBackendDataFrame}(object)

\S4method{precursorCharge}{MsBackendDataFrame}(object)

\S4method{precursorIntensity}{MsBackendDataFrame}(object)

\S4method{precursorMz}{MsBackendDataFrame}(object)

\S4method{rtime}{MsBackendDataFrame}(object)

\S4method{rtime}{MsBackendDataFrame}(object) <- value

\S4method{scanIndex}{MsBackendDataFrame}(object)

\S4method{selectSpectraVariables}{MsBackendDataFrame}(object,
  spectraVariables = spectraVariables(object))

\S4method{smoothed}{MsBackendDataFrame}(object)

\S4method{smoothed}{MsBackendDataFrame}(object) <- value

\S4method{spectraData}{MsBackendDataFrame}(object,
  columns = spectraVariables(object))

\S4method{spectraData}{MsBackendDataFrame}(object) <- value

\S4method{spectraNames}{MsBackendDataFrame}(object)

\S4method{spectraNames}{MsBackendDataFrame}(object) <- value

\S4method{spectraVariables}{MsBackendDataFrame}(object)

\S4method{tic}{MsBackendDataFrame}(object, initial = TRUE)

\S4method{$}{MsBackendDataFrame}(x, name)

\S4method{$}{MsBackendDataFrame}(x, name) <- value

\S4method{[}{MsBackendDataFrame,ANY}(x, i, j, ..., drop = FALSE)

\S4method{filterAcquisitionNum}{MsBackendDataFrame}(object,
  n = integer(), file = integer())

\S4method{filterEmptySpectra}{MsBackendDataFrame}(object)

\S4method{filterFile}{MsBackendDataFrame}(object, file = integer())

\S4method{backendInitialize}{MsBackendMzR}(object, files, spectraData, ...,
  BPPARAM = bpparam())

\S4method{show}{MsBackendMzR}(object)

\S4method{acquisitionNum}{MsBackendMzR}(object)

\S4method{centroided}{MsBackendMzR}(object)

\S4method{centroided}{MsBackendMzR}(object) <- value

\S4method{collisionEnergy}{MsBackendMzR}(object)

\S4method{collisionEnergy}{MsBackendMzR}(object) <- value

\S4method{fromFile}{MsBackendMzR}(object)

\S4method{intensity}{MsBackendMzR}(object)

\S4method{ionCount}{MsBackendMzR}(object)

\S4method{isCentroided}{MsBackendMzR}(object, ...)

\S4method{isEmpty}{MsBackendMzR}(x)

\S4method{msLevel}{MsBackendMzR}(object, ...)

\S4method{mz}{MsBackendMzR}(object)

\S4method{peaks}{MsBackendMzR}(object)

\S4method{peaksCount}{MsBackendMzR}(object)

\S4method{polarity}{MsBackendMzR}(object)

\S4method{polarity}{MsBackendMzR}(object) <- value

\S4method{precScanNum}{MsBackendMzR}(object)

\S4method{precursorCharge}{MsBackendMzR}(object)

\S4method{precursorIntensity}{MsBackendMzR}(object)

\S4method{precursorMz}{MsBackendMzR}(object)

\S4method{rtime}{MsBackendMzR}(object)

\S4method{rtime}{MsBackendMzR}(object) <- value

\S4method{scanIndex}{MsBackendMzR}(object)

\S4method{smoothed}{MsBackendMzR}(object)

\S4method{smoothed}{MsBackendMzR}(object) <- value

\S4method{spectraData}{MsBackendMzR}(object,
  columns = spectraVariables(object))

\S4method{spectraData}{MsBackendMzR}(object) <- value

\S4method{spectraNames}{MsBackendMzR}(object)

\S4method{spectraNames}{MsBackendMzR}(object) <- value

\S4method{spectraVariables}{MsBackendMzR}(object)

\S4method{tic}{MsBackendMzR}(object, initial = TRUE)

\S4method{$}{MsBackendMzR}(x, name) <- value

\S4method{show}{ProcessingStep}(object)

\S4method{show}{Spectra}(object)
}
\description{
For S4 methods that require a documentation entry but only clutter the index.
}
\keyword{internal}
