% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_aliases.R, R/AllGenerics.R,
%   R/MsBackendDataFrame.R, R/MsBackendHdf5Peaks.R, R/MsBackendMemory.R,
%   R/MsBackendMzR.R, R/Spectra.R
\docType{methods}
\name{hidden_aliases}
\alias{hidden_aliases}
\alias{[,MsBackendDataFrame-method}
\alias{ppm}
\alias{bin,numeric-method}
\alias{show,MsBackendDataFrame-method}
\alias{backendMerge,MsBackendDataFrame-method}
\alias{backendRequiredSpectraVariables,MsBackendDataFrame-method}
\alias{acquisitionNum,MsBackendDataFrame-method}
\alias{peaksData,MsBackendDataFrame-method}
\alias{centroided,MsBackendDataFrame-method}
\alias{centroided<-,MsBackendDataFrame-method}
\alias{collisionEnergy,MsBackendDataFrame-method}
\alias{collisionEnergy<-,MsBackendDataFrame-method}
\alias{dataOrigin,MsBackendDataFrame-method}
\alias{dataOrigin<-,MsBackendDataFrame-method}
\alias{dataStorage,MsBackendDataFrame-method}
\alias{dataStorage<-,MsBackendDataFrame-method}
\alias{extractByIndex,MsBackendDataFrame,ANY-method}
\alias{intensity,MsBackendDataFrame-method}
\alias{intensity<-,MsBackendDataFrame-method}
\alias{isEmpty,MsBackendDataFrame-method}
\alias{isolationWindowLowerMz,MsBackendDataFrame-method}
\alias{isolationWindowLowerMz<-,MsBackendDataFrame-method}
\alias{isolationWindowTargetMz,MsBackendDataFrame-method}
\alias{isolationWindowTargetMz<-,MsBackendDataFrame-method}
\alias{isolationWindowUpperMz,MsBackendDataFrame-method}
\alias{isolationWindowUpperMz<-,MsBackendDataFrame-method}
\alias{length,MsBackendDataFrame-method}
\alias{lengths,MsBackendDataFrame-method}
\alias{msLevel,MsBackendDataFrame-method}
\alias{msLevel<-,MsBackendDataFrame-method}
\alias{mz,MsBackendDataFrame-method}
\alias{mz<-,MsBackendDataFrame-method}
\alias{polarity,MsBackendDataFrame-method}
\alias{polarity<-,MsBackendDataFrame-method}
\alias{precScanNum,MsBackendDataFrame-method}
\alias{precursorCharge,MsBackendDataFrame-method}
\alias{precursorIntensity,MsBackendDataFrame-method}
\alias{precursorMz,MsBackendDataFrame-method}
\alias{peaksData<-,MsBackendDataFrame-method}
\alias{peaksVariables,MsBackendDataFrame-method}
\alias{rtime,MsBackendDataFrame-method}
\alias{rtime<-,MsBackendDataFrame-method}
\alias{scanIndex,MsBackendDataFrame-method}
\alias{selectSpectraVariables,MsBackendDataFrame-method}
\alias{smoothed,MsBackendDataFrame-method}
\alias{smoothed<-,MsBackendDataFrame-method}
\alias{spectraData,MsBackendDataFrame-method}
\alias{spectraData<-,MsBackendDataFrame-method}
\alias{spectraNames,MsBackendDataFrame-method}
\alias{spectraNames<-,MsBackendDataFrame-method}
\alias{spectraVariables,MsBackendDataFrame-method}
\alias{tic,MsBackendDataFrame-method}
\alias{$,MsBackendDataFrame-method}
\alias{$<-,MsBackendDataFrame-method}
\alias{cbind2,MsBackendDataFrame,dataframeOrDataFrameOrmatrix-method}
\alias{split,MsBackendDataFrame,ANY-method}
\alias{filterAcquisitionNum,MsBackendDataFrame-method}
\alias{backendRequiredSpectraVariables,MsBackendHdf5Peaks-method}
\alias{backendInitialize,MsBackendHdf5Peaks-method}
\alias{show,MsBackendHdf5Peaks-method}
\alias{peaksData,MsBackendHdf5Peaks-method}
\alias{intensity,MsBackendHdf5Peaks-method}
\alias{intensity<-,MsBackendHdf5Peaks-method}
\alias{ionCount,MsBackendHdf5Peaks-method}
\alias{isCentroided,MsBackendHdf5Peaks-method}
\alias{isEmpty,MsBackendHdf5Peaks-method}
\alias{lengths,MsBackendHdf5Peaks-method}
\alias{mz,MsBackendHdf5Peaks-method}
\alias{mz<-,MsBackendHdf5Peaks-method}
\alias{peaksData<-,MsBackendHdf5Peaks-method}
\alias{spectraData,MsBackendHdf5Peaks-method}
\alias{spectraData<-,MsBackendHdf5Peaks-method}
\alias{$<-,MsBackendHdf5Peaks-method}
\alias{[,MsBackendHdf5Peaks-method}
\alias{extractByIndex,MsBackendHdf5Peaks,ANY-method}
\alias{backendMerge,MsBackendHdf5Peaks-method}
\alias{show,MsBackendMemory-method}
\alias{backendMerge,MsBackendMemory-method}
\alias{backendRequiredSpectraVariables,MsBackendMemory-method}
\alias{acquisitionNum,MsBackendMemory-method}
\alias{centroided,MsBackendMemory-method}
\alias{centroided<-,MsBackendMemory-method}
\alias{collisionEnergy,MsBackendMemory-method}
\alias{collisionEnergy<-,MsBackendMemory-method}
\alias{dataOrigin,MsBackendMemory-method}
\alias{dataOrigin<-,MsBackendMemory-method}
\alias{dataStorage,MsBackendMemory-method}
\alias{dataStorage<-,MsBackendMemory-method}
\alias{extractByIndex,MsBackendMemory,ANY-method}
\alias{intensity,MsBackendMemory-method}
\alias{intensity<-,MsBackendMemory-method}
\alias{ionCount,MsBackendMemory-method}
\alias{isEmpty,MsBackendMemory-method}
\alias{isolationWindowLowerMz,MsBackendMemory-method}
\alias{isolationWindowLowerMz<-,MsBackendMemory-method}
\alias{isolationWindowTargetMz,MsBackendMemory-method}
\alias{isolationWindowTargetMz<-,MsBackendMemory-method}
\alias{isolationWindowUpperMz,MsBackendMemory-method}
\alias{isolationWindowUpperMz<-,MsBackendMemory-method}
\alias{length,MsBackendMemory-method}
\alias{lengths,MsBackendMemory-method}
\alias{msLevel,MsBackendMemory-method}
\alias{msLevel<-,MsBackendMemory-method}
\alias{mz,MsBackendMemory-method}
\alias{mz<-,MsBackendMemory-method}
\alias{peaksData,MsBackendMemory-method}
\alias{peaksData<-,MsBackendMemory-method}
\alias{polarity,MsBackendMemory-method}
\alias{polarity<-,MsBackendMemory-method}
\alias{precScanNum,MsBackendMemory-method}
\alias{precursorCharge,MsBackendMemory-method}
\alias{precursorIntensity,MsBackendMemory-method}
\alias{precursorMz,MsBackendMemory-method}
\alias{rtime,MsBackendMemory-method}
\alias{rtime<-,MsBackendMemory-method}
\alias{scanIndex,MsBackendMemory-method}
\alias{selectSpectraVariables,MsBackendMemory-method}
\alias{smoothed,MsBackendMemory-method}
\alias{smoothed<-,MsBackendMemory-method}
\alias{spectraData,MsBackendMemory-method}
\alias{spectraData<-,MsBackendMemory-method}
\alias{spectraNames,MsBackendMemory-method}
\alias{spectraNames<-,MsBackendMemory-method}
\alias{spectraVariables,MsBackendMemory-method}
\alias{peaksVariables,MsBackendMemory-method}
\alias{tic,MsBackendMemory-method}
\alias{$,MsBackendMemory-method}
\alias{$<-,MsBackendMemory-method}
\alias{[,MsBackendMemory-method}
\alias{cbind2,MsBackendMemory,dataframeOrDataFrameOrmatrix-method}
\alias{split,MsBackendMemory,ANY-method}
\alias{filterAcquisitionNum,MsBackendMemory-method}
\alias{longForm,MsBackendMemory-method}
\alias{backendRequiredSpectraVariables,MsBackendMzR-method}
\alias{backendInitialize,MsBackendMzR-method}
\alias{show,MsBackendMzR-method}
\alias{peaksData,MsBackendMzR-method}
\alias{intensity,MsBackendMzR-method}
\alias{intensity<-,MsBackendMzR-method}
\alias{ionCount,MsBackendMzR-method}
\alias{isCentroided,MsBackendMzR-method}
\alias{isEmpty,MsBackendMzR-method}
\alias{lengths,MsBackendMzR-method}
\alias{mz,MsBackendMzR-method}
\alias{mz<-,MsBackendMzR-method}
\alias{spectraData,MsBackendMzR-method}
\alias{spectraData<-,MsBackendMzR-method}
\alias{spectraNames,MsBackendMzR-method}
\alias{spectraNames<-,MsBackendMzR-method}
\alias{spectraVariables,MsBackendMzR-method}
\alias{$<-,MsBackendMzR-method}
\alias{export,MsBackendMzR-method}
\alias{show,Spectra-method}
\alias{combinePeaks,list-method}
\title{Internal page for hidden aliases}
\usage{
\S4method{bin}{numeric}(
  x,
  y,
  size = 1,
  breaks = seq(floor(min(y)), ceiling(max(y)), by = size),
  FUN = max,
  returnMids = TRUE,
  .check = TRUE
)

\S4method{show}{MsBackendDataFrame}(object)

\S4method{backendMerge}{MsBackendDataFrame}(object, ...)

\S4method{backendRequiredSpectraVariables}{MsBackendDataFrame}(object, ...)

\S4method{acquisitionNum}{MsBackendDataFrame}(object)

\S4method{peaksData}{MsBackendDataFrame}(object, columns = c("mz", "intensity"))

\S4method{centroided}{MsBackendDataFrame}(object)

\S4method{centroided}{MsBackendDataFrame}(object) <- value

\S4method{collisionEnergy}{MsBackendDataFrame}(object)

\S4method{collisionEnergy}{MsBackendDataFrame}(object) <- value

\S4method{dataOrigin}{MsBackendDataFrame}(object)

\S4method{dataOrigin}{MsBackendDataFrame}(object) <- value

\S4method{dataStorage}{MsBackendDataFrame}(object)

\S4method{dataStorage}{MsBackendDataFrame}(object) <- value

\S4method{extractByIndex}{MsBackendDataFrame,ANY}(object, i)

\S4method{intensity}{MsBackendDataFrame}(object)

\S4method{intensity}{MsBackendDataFrame}(object) <- value

\S4method{isEmpty}{MsBackendDataFrame}(x)

\S4method{isolationWindowLowerMz}{MsBackendDataFrame}(object)

\S4method{isolationWindowLowerMz}{MsBackendDataFrame}(object) <- value

\S4method{isolationWindowTargetMz}{MsBackendDataFrame}(object)

\S4method{isolationWindowTargetMz}{MsBackendDataFrame}(object) <- value

\S4method{isolationWindowUpperMz}{MsBackendDataFrame}(object)

\S4method{isolationWindowUpperMz}{MsBackendDataFrame}(object) <- value

\S4method{length}{MsBackendDataFrame}(x)

\S4method{lengths}{MsBackendDataFrame}(x, use.names = FALSE)

\S4method{msLevel}{MsBackendDataFrame}(object, ...)

\S4method{msLevel}{MsBackendDataFrame}(object) <- value

\S4method{mz}{MsBackendDataFrame}(object)

\S4method{mz}{MsBackendDataFrame}(object) <- value

\S4method{polarity}{MsBackendDataFrame}(object)

\S4method{polarity}{MsBackendDataFrame}(object) <- value

\S4method{precScanNum}{MsBackendDataFrame}(object)

\S4method{precursorCharge}{MsBackendDataFrame}(object)

\S4method{precursorIntensity}{MsBackendDataFrame}(object)

\S4method{precursorMz}{MsBackendDataFrame}(object)

\S4method{peaksData}{MsBackendDataFrame}(object) <- value

\S4method{peaksVariables}{MsBackendDataFrame}(object)

\S4method{rtime}{MsBackendDataFrame}(object)

\S4method{rtime}{MsBackendDataFrame}(object) <- value

\S4method{scanIndex}{MsBackendDataFrame}(object)

\S4method{selectSpectraVariables}{MsBackendDataFrame}(object, spectraVariables = spectraVariables(object))

\S4method{smoothed}{MsBackendDataFrame}(object)

\S4method{smoothed}{MsBackendDataFrame}(object) <- value

\S4method{spectraData}{MsBackendDataFrame}(object, columns = spectraVariables(object))

\S4method{spectraData}{MsBackendDataFrame}(object) <- value

\S4method{spectraNames}{MsBackendDataFrame}(object)

\S4method{spectraNames}{MsBackendDataFrame}(object) <- value

\S4method{spectraVariables}{MsBackendDataFrame}(object)

\S4method{tic}{MsBackendDataFrame}(object, initial = TRUE)

\S4method{$}{MsBackendDataFrame}(x, name)

\S4method{$}{MsBackendDataFrame}(x, name) <- value

\S4method{[}{MsBackendDataFrame}(x, i, j, ..., drop = FALSE)

\S4method{cbind2}{MsBackendDataFrame,dataframeOrDataFrameOrmatrix}(x, y = data.frame(), ...)

\S4method{split}{MsBackendDataFrame,ANY}(x, f, drop = FALSE, ...)

\S4method{filterAcquisitionNum}{MsBackendDataFrame}(
  object,
  n = integer(),
  dataStorage = character(),
  dataOrigin = character()
)

\S4method{backendRequiredSpectraVariables}{MsBackendHdf5Peaks}(object, ...)

\S4method{backendInitialize}{MsBackendHdf5Peaks}(
  object,
  files = character(),
  data = DataFrame(),
  hdf5path = character(),
  ...,
  BPPARAM = bpparam()
)

\S4method{show}{MsBackendHdf5Peaks}(object)

\S4method{peaksData}{MsBackendHdf5Peaks}(object, columns = peaksVariables(object))

\S4method{intensity}{MsBackendHdf5Peaks}(object)

\S4method{intensity}{MsBackendHdf5Peaks}(object) <- value

\S4method{ionCount}{MsBackendHdf5Peaks}(object)

\S4method{isCentroided}{MsBackendHdf5Peaks}(object, ...)

\S4method{isEmpty}{MsBackendHdf5Peaks}(x)

\S4method{lengths}{MsBackendHdf5Peaks}(x, use.names = FALSE)

\S4method{mz}{MsBackendHdf5Peaks}(object)

\S4method{mz}{MsBackendHdf5Peaks}(object) <- value

\S4method{peaksData}{MsBackendHdf5Peaks}(object) <- value

\S4method{spectraData}{MsBackendHdf5Peaks}(object, columns = spectraVariables(object))

\S4method{spectraData}{MsBackendHdf5Peaks}(object) <- value

\S4method{$}{MsBackendHdf5Peaks}(x, name) <- value

\S4method{[}{MsBackendHdf5Peaks}(x, i, j, ..., drop = FALSE)

\S4method{extractByIndex}{MsBackendHdf5Peaks,ANY}(object, i)

\S4method{backendMerge}{MsBackendHdf5Peaks}(object, ...)

\S4method{show}{MsBackendMemory}(object)

\S4method{backendMerge}{MsBackendMemory}(object, ...)

\S4method{backendRequiredSpectraVariables}{MsBackendMemory}(object, ...)

\S4method{acquisitionNum}{MsBackendMemory}(object)

\S4method{centroided}{MsBackendMemory}(object)

\S4method{centroided}{MsBackendMemory}(object) <- value

\S4method{collisionEnergy}{MsBackendMemory}(object)

\S4method{collisionEnergy}{MsBackendMemory}(object) <- value

\S4method{dataOrigin}{MsBackendMemory}(object)

\S4method{dataOrigin}{MsBackendMemory}(object) <- value

\S4method{dataStorage}{MsBackendMemory}(object)

\S4method{dataStorage}{MsBackendMemory}(object) <- value

\S4method{extractByIndex}{MsBackendMemory,ANY}(object, i)

\S4method{intensity}{MsBackendMemory}(object)

\S4method{intensity}{MsBackendMemory}(object) <- value

\S4method{ionCount}{MsBackendMemory}(object)

\S4method{isEmpty}{MsBackendMemory}(x)

\S4method{isolationWindowLowerMz}{MsBackendMemory}(object)

\S4method{isolationWindowLowerMz}{MsBackendMemory}(object) <- value

\S4method{isolationWindowTargetMz}{MsBackendMemory}(object)

\S4method{isolationWindowTargetMz}{MsBackendMemory}(object) <- value

\S4method{isolationWindowUpperMz}{MsBackendMemory}(object)

\S4method{isolationWindowUpperMz}{MsBackendMemory}(object) <- value

\S4method{length}{MsBackendMemory}(x)

\S4method{lengths}{MsBackendMemory}(x, use.names = FALSE)

\S4method{msLevel}{MsBackendMemory}(object, ...)

\S4method{msLevel}{MsBackendMemory}(object) <- value

\S4method{mz}{MsBackendMemory}(object)

\S4method{mz}{MsBackendMemory}(object) <- value

\S4method{peaksData}{MsBackendMemory}(object, columns = c("mz", "intensity"))

\S4method{peaksData}{MsBackendMemory}(object) <- value

\S4method{polarity}{MsBackendMemory}(object)

\S4method{polarity}{MsBackendMemory}(object) <- value

\S4method{precScanNum}{MsBackendMemory}(object)

\S4method{precursorCharge}{MsBackendMemory}(object)

\S4method{precursorIntensity}{MsBackendMemory}(object)

\S4method{precursorMz}{MsBackendMemory}(object)

\S4method{rtime}{MsBackendMemory}(object)

\S4method{rtime}{MsBackendMemory}(object) <- value

\S4method{scanIndex}{MsBackendMemory}(object)

\S4method{selectSpectraVariables}{MsBackendMemory}(object, spectraVariables = spectraVariables(object))

\S4method{smoothed}{MsBackendMemory}(object)

\S4method{smoothed}{MsBackendMemory}(object) <- value

\S4method{spectraData}{MsBackendMemory}(object, columns = spectraVariables(object))

\S4method{spectraData}{MsBackendMemory}(object) <- value

\S4method{spectraNames}{MsBackendMemory}(object)

\S4method{spectraNames}{MsBackendMemory}(object) <- value

\S4method{spectraVariables}{MsBackendMemory}(object)

\S4method{peaksVariables}{MsBackendMemory}(object)

\S4method{tic}{MsBackendMemory}(object, initial = TRUE)

\S4method{$}{MsBackendMemory}(x, name)

\S4method{$}{MsBackendMemory}(x, name) <- value

\S4method{[}{MsBackendMemory}(x, i, j, ..., drop = FALSE)

\S4method{cbind2}{MsBackendMemory,dataframeOrDataFrameOrmatrix}(x, y = data.frame(), ...)

\S4method{split}{MsBackendMemory,ANY}(x, f, drop = FALSE, ...)

\S4method{filterAcquisitionNum}{MsBackendMemory}(
  object,
  n = integer(),
  dataStorage = character(),
  dataOrigin = character()
)

\S4method{longForm}{MsBackendMemory}(object, columns = spectraVariables(object))

\S4method{backendRequiredSpectraVariables}{MsBackendMzR}(object, ...)

\S4method{backendInitialize}{MsBackendMzR}(object, files, ..., BPPARAM = bpparam())

\S4method{show}{MsBackendMzR}(object)

\S4method{peaksData}{MsBackendMzR}(object, columns = peaksVariables(object))

\S4method{intensity}{MsBackendMzR}(object)

\S4method{intensity}{MsBackendMzR}(object) <- value

\S4method{ionCount}{MsBackendMzR}(object)

\S4method{isCentroided}{MsBackendMzR}(object, ...)

\S4method{isEmpty}{MsBackendMzR}(x)

\S4method{lengths}{MsBackendMzR}(x, use.names = FALSE)

\S4method{mz}{MsBackendMzR}(object)

\S4method{mz}{MsBackendMzR}(object) <- value

\S4method{spectraData}{MsBackendMzR}(object, columns = spectraVariables(object))

\S4method{spectraData}{MsBackendMzR}(object) <- value

\S4method{spectraNames}{MsBackendMzR}(object)

\S4method{spectraNames}{MsBackendMzR}(object) <- value

\S4method{spectraVariables}{MsBackendMzR}(object)

\S4method{$}{MsBackendMzR}(x, name) <- value

\S4method{export}{MsBackendMzR}(
  object,
  x,
  file = tempfile(),
  format = c("mzML", "mzXML"),
  copy = FALSE,
  BPPARAM = bpparam()
)

\S4method{show}{Spectra}(object)

\S4method{combinePeaks}{list}(object, ...)
}
\value{
Not applicable
}
\description{
For S4 methods that require a documentation entry but only clutter the index.
}
\note{
: this replaces \strong{all} the data in the backend.
}
\keyword{internal}
