% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_aliases.R, R/AllGenerics.R,
%   R/MsBackendDataFrame.R, R/MsBackendMzR.R, R/ProcessingStep.R, R/Spectra.R
\docType{methods}
\name{hidden_aliases}
\alias{hidden_aliases}
\alias{acquisitionNum}
\alias{backendInitialize}
\alias{centroided}
\alias{centroided<-}
\alias{collisionEnergy}
\alias{collisionEnergy<-}
\alias{fileNames}
\alias{fromFile}
\alias{ionCount}
\alias{isReadOnly}
\alias{msLevel}
\alias{peaksCount}
\alias{polarity}
\alias{polarity<-}
\alias{precScanNum}
\alias{precursorCharge}
\alias{precursorIntensity}
\alias{precursorMz}
\alias{rtime<-}
\alias{scanIndex}
\alias{smoothed}
\alias{smoothed<-}
\alias{spectraData}
\alias{spectraData<-}
\alias{spectraNames}
\alias{spectraNames<-}
\alias{spectraVariables}
\alias{show,MsBackendDataFrame-method}
\alias{backendInitialize,MsBackendDataFrame-method}
\alias{acquisitionNum,MsBackendDataFrame-method}
\alias{centroided,MsBackendDataFrame-method}
\alias{centroided<-,MsBackendDataFrame-method}
\alias{collisionEnergy,MsBackendDataFrame-method}
\alias{collisionEnergy<-,MsBackendDataFrame-method}
\alias{fromFile,MsBackendDataFrame-method}
\alias{intensity,MsBackendDataFrame-method}
\alias{ionCount,MsBackendDataFrame-method}
\alias{isCentroided,MsBackendDataFrame-method}
\alias{isEmpty,MsBackendDataFrame-method}
\alias{length,MsBackendDataFrame-method}
\alias{msLevel,MsBackendDataFrame-method}
\alias{mz,MsBackendDataFrame-method}
\alias{peaks,MsBackendDataFrame-method}
\alias{peaksCount,MsBackendDataFrame-method}
\alias{polarity,MsBackendDataFrame-method}
\alias{polarity<-,MsBackendDataFrame-method}
\alias{precScanNum,MsBackendDataFrame-method}
\alias{precursorCharge,MsBackendDataFrame-method}
\alias{precursorIntensity,MsBackendDataFrame-method}
\alias{precursorMz,MsBackendDataFrame-method}
\alias{rtime,MsBackendDataFrame-method}
\alias{rtime<-,MsBackendDataFrame-method}
\alias{scanIndex,MsBackendDataFrame-method}
\alias{smoothed,MsBackendDataFrame-method}
\alias{smoothed<-,MsBackendDataFrame-method}
\alias{spectraData,MsBackendDataFrame-method}
\alias{spectraData<-,MsBackendDataFrame-method}
\alias{spectraNames,MsBackendDataFrame-method}
\alias{spectraNames<-,MsBackendDataFrame-method}
\alias{spectraVariables,MsBackendDataFrame-method}
\alias{tic,MsBackendDataFrame-method}
\alias{[,MsBackendDataFrame,ANY,ANY,ANY-method}
\alias{backendInitialize,MsBackendMzR-method}
\alias{acquisitionNum,MsBackendMzR-method}
\alias{centroided,MsBackendMzR-method}
\alias{centroided<-,MsBackendMzR-method}
\alias{collisionEnergy,MsBackendMzR-method}
\alias{collisionEnergy<-,MsBackendMzR-method}
\alias{fromFile,MsBackendMzR-method}
\alias{msLevel,MsBackendMzR-method}
\alias{polarity,MsBackendMzR-method}
\alias{polarity<-,MsBackendMzR-method}
\alias{precScanNum,MsBackendMzR-method}
\alias{precursorCharge,MsBackendMzR-method}
\alias{precursorIntensity,MsBackendMzR-method}
\alias{precursorMz,MsBackendMzR-method}
\alias{rtime,MsBackendMzR-method}
\alias{rtime<-,MsBackendMzR-method}
\alias{scanIndex,MsBackendMzR-method}
\alias{smoothed,MsBackendMzR-method}
\alias{smoothed<-,MsBackendMzR-method}
\alias{spectraNames,MsBackendMzR-method}
\alias{spectraVariables,MsBackendMzR-method}
\alias{ProcessingStep-class}
\alias{show,ProcessingStep-method}
\alias{show,Spectra-method}
\title{Internal page for hidden aliases}
\usage{
acquisitionNum(object, ...)

backendInitialize(object, files, spectraData, ...)

centroided(object, ...)

centroided(object) <- value

collisionEnergy(object, ...)

collisionEnergy(object) <- value

fileNames(object, ...)

fromFile(object, ...)

ionCount(object, ...)

isReadOnly(object, ...)

msLevel(object, ...)

peaksCount(object, ...)

polarity(object, ...)

polarity(object) <- value

precScanNum(object, ...)

precursorCharge(object, ...)

precursorIntensity(object, ...)

precursorMz(object, ...)

rtime(object) <- value

scanIndex(object, ...)

smoothed(object, ...)

smoothed(object) <- value

spectraData(object, ...)

spectraData(object) <- value

spectraNames(object, ...)

spectraNames(object) <- value

spectraVariables(object, ...)

\S4method{show}{MsBackendDataFrame}(object)

\S4method{backendInitialize}{MsBackendDataFrame}(object, files,
  spectraData, ...)

\S4method{acquisitionNum}{MsBackendDataFrame}(object)

\S4method{centroided}{MsBackendDataFrame}(object)

\S4method{centroided}{MsBackendDataFrame}(object) <- value

\S4method{collisionEnergy}{MsBackendDataFrame}(object)

\S4method{collisionEnergy}{MsBackendDataFrame}(object) <- value

\S4method{fromFile}{MsBackendDataFrame}(object)

\S4method{intensity}{MsBackendDataFrame}(object)

\S4method{ionCount}{MsBackendDataFrame}(object)

\S4method{isCentroided}{MsBackendDataFrame}(object, ...)

\S4method{isEmpty}{MsBackendDataFrame}(x)

\S4method{length}{MsBackendDataFrame}(x)

\S4method{msLevel}{MsBackendDataFrame}(object, ...)

\S4method{mz}{MsBackendDataFrame}(object)

\S4method{peaks}{MsBackendDataFrame}(object)

\S4method{peaksCount}{MsBackendDataFrame}(object)

\S4method{polarity}{MsBackendDataFrame}(object)

\S4method{polarity}{MsBackendDataFrame}(object) <- value

\S4method{precScanNum}{MsBackendDataFrame}(object)

\S4method{precursorCharge}{MsBackendDataFrame}(object)

\S4method{precursorIntensity}{MsBackendDataFrame}(object)

\S4method{precursorMz}{MsBackendDataFrame}(object)

\S4method{rtime}{MsBackendDataFrame}(object)

\S4method{rtime}{MsBackendDataFrame}(object) <- value

\S4method{scanIndex}{MsBackendDataFrame}(object)

\S4method{smoothed}{MsBackendDataFrame}(object)

\S4method{smoothed}{MsBackendDataFrame}(object) <- value

\S4method{spectraData}{MsBackendDataFrame}(object,
  columns = spectraVariables(object))

\S4method{spectraData}{MsBackendDataFrame}(object) <- value

\S4method{spectraNames}{MsBackendDataFrame}(object)

\S4method{spectraNames}{MsBackendDataFrame}(object) <- value

\S4method{spectraVariables}{MsBackendDataFrame}(object)

\S4method{tic}{MsBackendDataFrame}(object, initial = TRUE)

\S4method{[}{MsBackendDataFrame,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{backendInitialize}{MsBackendMzR}(object, files, spectraData, ...,
  BPPARAM = bpparam())

\S4method{acquisitionNum}{MsBackendMzR}(object)

\S4method{centroided}{MsBackendMzR}(object)

\S4method{centroided}{MsBackendMzR}(object) <- value

\S4method{collisionEnergy}{MsBackendMzR}(object)

\S4method{collisionEnergy}{MsBackendMzR}(object) <- value

\S4method{fromFile}{MsBackendMzR}(object)

\S4method{msLevel}{MsBackendMzR}(object, ...)

\S4method{polarity}{MsBackendMzR}(object)

\S4method{polarity}{MsBackendMzR}(object) <- value

\S4method{precScanNum}{MsBackendMzR}(object)

\S4method{precursorCharge}{MsBackendMzR}(object)

\S4method{precursorIntensity}{MsBackendMzR}(object)

\S4method{precursorMz}{MsBackendMzR}(object)

\S4method{rtime}{MsBackendMzR}(object)

\S4method{rtime}{MsBackendMzR}(object) <- value

\S4method{scanIndex}{MsBackendMzR}(object)

\S4method{smoothed}{MsBackendMzR}(object)

\S4method{smoothed}{MsBackendMzR}(object) <- value

\S4method{spectraNames}{MsBackendMzR}(object)

\S4method{spectraVariables}{MsBackendMzR}(object)

\S4method{show}{ProcessingStep}(object)

\S4method{show}{Spectra}(object)
}
\description{
For S4 methods that require a documentation entry but only clutter the index.
}
\keyword{internal}
