% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsBackend-functions.R, R/MsBackend.R
\name{MsBackend}
\alias{MsBackend}
\title{Mass spectrometry data backends}
\usage{
MsBackend()
}
\description{
\linkS4class{MsBackend} objects provide access to mass spectrometry data. Such
backends can be generally classidied into \emph{in-memory} and \emph{on-disk} backends.
In-memory backends, such as the base \code{MsBackend}, keep all (spectra) data in
memory ensuring fast data access. On-disk backends keep only part of the
data in memory retrieving the remaining data (mostly m/z and intensity
values) on-demand from disk.

Available backends for \code{\link[=Spectra]{Spectra()}} objects are listed below. The last section
provides information for the development of new backends.
}
\section{\code{MsBackend}}{


The base backend class, \code{MsBackend}, keeps all spectra data in a \code{DataFrame}.
}

\section{Implementation notes for new backend classes}{


New backend classes should extend the base \code{MsBackend} class.

TODO: add description of required methods etc.
}

\author{
Johannes Rainer, Sebastian Gibb, Laurent Gatto
}
