% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotting-functions.R
\name{spectra-ggplotting}
\alias{spectra-ggplotting}
\alias{ggplotSpectraMirror}
\alias{ggplotSpectra}
\alias{ggplotSpectraOverlay}
\alias{ggplotSpectraMirror,Spectra-method}
\title{Plotting Spectra with ggplot}
\usage{
ggplotSpectra(
  x,
  xlab = "m/z",
  ylab = "intensity",
  xlim = numeric(),
  ylim = numeric(),
  main = character(),
  col = "#00000080",
  labels = character(),
  labelCol = col,
  labelSize = 5,
  labelAngle = 0,
  labelVjust = -0.2,
  labelHjust = 0.5,
  asp = 0.5,
  axes = TRUE,
  frame.plot = axes,
  interactive = FALSE,
  interactive_labels = c("rtime", "mz", "intensity", "msLevel")
)

ggplotSpectraOverlay(
  x,
  xlab = "m/z",
  ylab = "intensity",
  xlim = numeric(),
  ylim = numeric(),
  main = paste(length(x), "spectra"),
  col = "#00000080",
  labels = character(),
  labelCol = col,
  labelSize = 5,
  labelAngle = 0,
  labelVjust = -0.2,
  labelHjust = 0.5,
  asp = 0.5,
  axes = TRUE,
  frame.plot = axes,
  interactive = FALSE,
  interactive_labels = c("rtime", "mz", "intensity", "msLevel")
)

\S4method{ggplotSpectraMirror}{Spectra}(
  x,
  y,
  xlab = "m/z",
  ylab = "intensity",
  xlim = numeric(),
  ylim = numeric(),
  main = character(),
  col = "#00000080",
  labels = character(),
  labelCol = col,
  labelSize = 5,
  labelAngle = 0,
  labelVjust = -0.2,
  labelHjust = 0.5,
  axes = TRUE,
  frame.plot = axes,
  ppm = 20,
  tolerance = 0,
  matchCol = "#80B1D3",
  matchCex = 5,
  matchPch = 16,
  matchLwd = 0.5,
  asp = 0.5,
  interactive = FALSE,
  interactive_labels = c("rtime", "mz", "intensity", "msLevel")
)
}
\arguments{
\item{x}{a \code{\link[=Spectra]{Spectra()}} object. For \code{ggplotSpectraMirror()} it has to be an
object of length 2.}

\item{xlab}{\code{character(1)} with the label for the x-axis (by default
\code{xlab = "m/z"}).}

\item{ylab}{\code{character(1)} with the label for the y-axis (by default
\code{ylab = "intensity"}).}

\item{xlim}{\code{numeric(2)} defining the x-axis limits. The range of m/z values
are used by default.}

\item{ylim}{\code{numeric(2)} defining the y-axis limits. The range of intensity
values are used by default.}

\item{main}{\code{character(1)} with the title for the plot. By default the
spectrum's MS level and retention time (in seconds) is used.}

\item{col}{color to be used to draw the peaks. Should be either of length 1,
or equal to the number of spectra (to plot each spectrum in a different
color) or be a \code{list} with colors for each individual peak in each
spectrum.}

\item{labels}{allows to specify a label for each peak. Needs to be a \code{list()}
with length equal to the number of spectra (each element of the list
being a \code{character()} with length equal to the number of peaks for that
spectrum), or, ideally, a \code{function} that uses one of the \code{Spectra}'s
variables (see examples below). \code{ggplotSpectraMirror()} supports only
\code{labels} of type \emph{function}.}

\item{labelCol}{color for the label(s).}

\item{labelSize}{size of the label(s).}

\item{labelAngle}{\code{numeric(1)} defining the rotation of the label. See
parameter \code{angle} in \emph{Aesthetics} of \code{\link[ggplot2:geom_text]{ggplot2::geom_text()}}.}

\item{labelVjust}{vertical justification of the label(s).}

\item{labelHjust}{horizontal justification of the label(s).}

\item{asp}{aspect ratio of the plot(s) Default: \code{1/2}.}

\item{axes}{\code{logical(1)} whether (x and y) axes should be drawn.}

\item{frame.plot}{\code{logical(1)} whether a box should be drawn around the
plotting area.}

\item{interactive}{\code{logical(1)} return the interactive ggplot based on
ggiraph.}

\item{interactive_labels}{\code{character} vector with the column names to show
in the interactive label(s).}

\item{y}{for \code{ggplotSpectraMirror()}: \code{Spectra} object of length 1 against
which \code{x} should be plotted against.}

\item{ppm}{for \code{ggplotSpectraMirror()}: m/z relative acceptable difference
(in ppm) for peaks to be considered matching (see \code{\link[MsCoreUtils:common]{MsCoreUtils::common()}}
for more details).}

\item{tolerance}{for \code{ggplotSpectraMirror()}: absolute acceptable difference
of m/z values for peaks to be considered matching (see
\code{\link[MsCoreUtils:common]{MsCoreUtils::common()}} for more details).}

\item{matchCol}{for \code{ggplotSpectraMirror()}: color for matching peaks.}

\item{matchCex}{for \code{ggplotSpectraMirror()}: point size to draw matching
peaks.}

\item{matchPch}{for \code{ggplotSpectraMirror()}: point character to label
matching peaks. Defaults to \code{matchPch = 16}, set to \code{matchPch = NA} to
disable.}

\item{matchLwd}{for \code{ggplotSpectraMirror()}: line width to draw matching
peaks.}
}
\value{
These functions create a ggplot.
}
\description{
\code{\link[=Spectra]{Spectra()}} can be plotted with one of the following functions
\itemize{
\item \code{ggplotSpectra()}: plots each spectrum in its separate plot by splitting
the plot area into as many panels as there are spectra.
\item \code{ggplotSpectraOverlay()}: plots all spectra in \code{x} \strong{into the same} plot
(as an overlay).
\item \code{ggplotSpectraMirror()}: plots a pair of spectra as a \emph{mirror plot}.
Parameters \code{x} and \code{y} both have to be a \code{Spectra} of length 1. Matching
peaks (considering \code{ppm} and \code{tolerance}) are highlighted. See
\code{\link[MsCoreUtils:common]{MsCoreUtils::common()}} for details on peak matching. Parameters
\code{matchCol}, \code{matchLty}, \code{matchLwd} and \code{matchPch} allow to customize
how matching peaks are indicated.
}
}
\examples{

ints <- list(c(4.3412, 12, 8, 34, 23.4),
    c(8, 25, 16, 32))
mzs <- list(c(13.453421, 43.433122, 46.6653553, 129.111212, 322.24432),
    c(13.452, 43.5122, 129.112, 322.245))

df <- DataFrame(msLevel = c(1L, 1L), rtime = c(123.12, 124))
df$mz <- mzs
df$intensity <- ints
sp <- Spectra(df)

#### --------------------------------------------- ####
##                   ggplotSpectra                   ##

## Plot one spectrum.
ggplotSpectra(sp[1])

## Plot both spectra.
ggplotSpectra(sp)

## Define a color for each peak in each spectrum.
ggplotSpectra(sp, col = list(c(1, 2, 3, 4, 5), 1:4))

## Color peaks from each spectrum in different colors.
ggplotSpectra(sp, col = c("green", "blue"))

## Label each peak with its m/z.
ggplotSpectra(sp, labels = function(z) lapply(mz(z), format, digits = 4))

## Rotate the labels.
ggplotSpectra(sp, labels = function(z) lapply(mz(z), format, digits = 4),
    labelAngle = -30, labelHjust = 0)

## Add a custom annotation for each peak.
sp$label <- list(c("", "A", "B", "C", "D"),
    c("Frodo", "Bilbo", "Peregrin", "Samwise"))

## Plot each peak in a different color
ggplotSpectra(sp, labels = sp$label,
    col = list(1:5, 1:4))

## Plot a single spectrum specifying the label.
ggplotSpectra(sp[2], labels = list(c("A", "B", "C", "D")))


#### --------------------------------------------- ####
##                ggplotSpectraOverlay               ##

## Plot both spectra overlaying.
ggplotSpectraOverlay(sp)

## Use a different color for each spectrum.
ggplotSpectraOverlay(sp, col = c("#ff000080", "#0000ff80"))

## Label also the peaks with their m/z if their intensity is above 15.
ggplotSpectraOverlay(sp, col = c("#ff000080", "#0000ff80"),
labels = function(z) {
    lapply(seq_along(mz(z)), function(i) {
            lbls <- format(mz(z)[[i]], digits = 4)
            lbls[intensity(z)[[i]] <= 15] <- ""
            lbls
     })
 }) + ggplot2::geom_hline(yintercept = 15, linetype = 2)

## Use different asp values
ggplotSpectra(sp, asp = 1/3)
ggplotSpectra(sp, asp = 2/1)

#### --------------------------------------------- ####
##                ggplotSpectraMirror                ##

## Plot two spectra against each other.
ggplotSpectraMirror(sp[1], sp[2])

## Label the peaks with their m/z
ggplotSpectraMirror(sp[1], sp[2],
    labels = function(z) list(format(mz(z)[[1L]], digits = 3)),
    labelAngle = -30, labelHjust = 0)

## The same ggplot with a tolerance of 0.1 and using a different color to
## highlight matching peaks
ggplotSpectraMirror(sp[1], sp[2],
    labels = function(z) list(format(mz(z)[[1L]], digits = 3)),
    labelAngle = -30, labelHjust = 0, tolerance = 0.1,
    matchCol = "#ff000080", matchLwd = 2, matchCex = 8)
}
\author{
Gabriele Tomè, Johannes Rainer
}
