% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra.R
\docType{methods}
\name{Spectra}
\alias{Spectra}
\alias{Spectra-class}
\alias{Spectra,DataFrame-method}
\alias{Spectra,missing-method}
\alias{length,Spectra-method}
\alias{msLevel,Spectra-method}
\title{The Spectra class to manage and access MS data}
\usage{
\S4method{Spectra}{DataFrame}(object, processingQueue = list(),
  metadata = list(), ..., backend = MsBackendDataFrame(),
  BPPARAM = bpparam())

\S4method{Spectra}{missing}(object, processingQueue = list(),
  metadata = list(), ..., backend = MsBackendDataFrame(),
  BPPARAM = bpparam())

\S4method{length}{Spectra}(x)

\S4method{msLevel}{Spectra}(object)
}
\value{
See individual method description for the return value.

TODO @jo: add all parameters (in alphabetic order).
}
\description{
The \code{Spectra} class encapsules spectral mass spectrometry data and
related metadata.

It supports multiple data backends, e.g. in-memory (\code{\link[=MsBackendDataFrame]{MsBackendDataFrame()}}),
on-disk as mzML (\code{\link[=MsBackendMzR]{MsBackendMzR()}}).
}
\details{
The \code{Spectra} class uses by default a lazy data manipulation strategy,
i.e. data manipulations such as performed with \code{removePeaks} are not applied
immediately to the data, but applied on-the-fly to the spectrum data once it
is retrieved.
}
\section{Creation of objects, conversion and changing the backend}{


\code{Spectra} classes can be created with the \code{Spectra} constructor function
which supports the following formats:
\itemize{
\item parameter \code{object} is a \code{DataFrame} containing the spectrum data. The
provided \code{backend} (by default a \linkS4class{MsBackendDataFrame}) will be
initialized with that data.
\item parameter \code{object} is missing, in which case it is supposed that the data
is provided by the \linkS4class{MsBackend} class passed along with the \code{backend}
argument.
}

\code{Spectra} classes are usually created with the \code{readSpectra}
function that reads general spectrum metadata information from the  mass
spectrometry data files.

The \linkS4class{MsBackend} can be changed with the \code{setBackend} function by
specifying the new \linkS4class{MsBackend} with the \code{backend} parameter. See
examples for more details.
}

\examples{

## Create a Spectra providing a `DataFrame` containing the spectrum data.

spd <- DataFrame(msLevel = c(1L, 2L), rtime = c(1.1, 1.2))
spd$mz <- list(c(100, 103.2, 104.3, 106.5), c(45.6, 120.4, 190.2))
spd$intensity <- list(c(200, 400, 34.2, 17), c(12.3, 15.2, 6.8))

data <- Spectra(spd)
data

msLevel(data)
}
\author{
Sebastian Gibb, Johannes Rainer
}
