% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-functions.R, R/Spectra.R
\docType{methods}
\name{Spectra}
\alias{Spectra}
\alias{Spectra-class}
\alias{length,Spectra-method}
\alias{msLevel,Spectra-method}
\title{The Spectra class to manage and access MS data}
\usage{
Spectra(spectraData = DataFrame(), msLevel, rt, acquisitionNum,
  scanIndex, fromFile, centroided, smoothed, polarity, precScanNum,
  precursorMz, precursorIntensity, precursorCharge, collisionEnergy, mz,
  intensity, backend = MsBackend(), processingQueue = list(),
  metadata = list())

\S4method{length}{Spectra}(x)

\S4method{msLevel}{Spectra}(object)
}
\value{
See individual method description for the return value.
}
\description{
The \code{Spectra} class encapsules spectral mass spectrometry data and
related metadata.

It supports multiple data backends, e.g. in-memory (\linkS4class{MsBackend}),
on-disk as mzML (\linkS4class{MsBackendMzR}) or HDF5 (\linkS4class{MsBackendHdf5}).
}
\details{
The \code{Spectra} class uses by default a lazy data manipulation strategy,
i.e. data manipulations such as performed with \code{removePeaks} are not applied
immediately to the data, but applied on-the-fly to the spectrum data once it
is retrieved.
}
\section{Creation of objects, conversion and changing the backend}{


TODO @jo: update with the \code{Spectra} constructor.

\code{Spectra} classes are usually created with the \code{readSpectra}
function that reads general spectrum metadata information from the  mass
spectrometry data files.

Alternatively it is possible to create a new object from a list of \code{Spectrum}
objects using the \code{Spectra} function. Additional spectrum metadata
columns can be provided with the \code{spectraData} argument, sample annotations
with the \code{sampleData} argument and arbitrary metadata with the \code{metadata}
argument. Note that objects created with the \code{Spectra} constructor
function can not use the \code{BackendMzR} as backend.

\code{Spectra} objects can be converted to a \code{list} or
\link[S4Vectors:List-class]{S4Vectors::List} of \code{Spectrum} objects with the \code{as(object, "list")}
and \code{as(object, "List")} function, respectively.

The \linkS4class{MsBackend} can be changed with the \code{setBackend} function by
specifying the new \linkS4class{MsBackend} with the \code{backend} parameter. See
examples for more details.
}

\examples{

## Create an Spectra from a list of Spectrum objects
}
\author{
Sebastian Gibb, Johannes Rainer
}
