% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-spectra.R
\docType{data}
\name{spectra}
\alias{spectra}
\title{Reflectance spectra}
\description{
Individual biological spectra, binned at 1 nm intervals from
 300 - 700 nm, beginning with a wavelength \code{'wl'} column.
}
\author{
Thomas White \email{thomas.white026@gmail.com}
}
\keyword{data}
